% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{chimeraUchime}
\alias{chimeraUchime}
\title{chimeraUchime}
\usage{
chimeraUchime(dataset, options)
}
\arguments{
\item{dataset}{R6 'sequence_dataset' object containing sequence data}

\item{options}{list containing parameter options}
}
\description{
detects and removes chimeras from your data using a denovo
approach.
}
\seealso{
\code{\link[=chimera_uchime]{chimera_uchime()}}
}
