% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_name}
\alias{extract_name}
\title{extract_name}
\usage{
extract_name(line)
}
\arguments{
\item{line, }{String containing name data}
}
\value{
A string
}
\description{
extracts name from fasta file entry
}
\examples{
(extract_name("mySeq This is my comment"))

}
