% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_uchime_version.R
\name{get_uchime_version}
\alias{get_uchime_version}
\title{get_uchime_version}
\usage{
get_uchime_version()
}
\value{
String, containing the version of uchime
}
\description{
Outputs the version of uchime, written by Robert C. Edgar,
modified for use within an R package.
}
\examples{

get_uchime_version()

}
\author{
Sarah Westcott, \email{swestcot@umich.edu}
}
