% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mothur2_example.R
\name{mothur2_example}
\alias{mothur2_example}
\title{Get path to mothur2 example}
\usage{
mothur2_example(file = NULL)
}
\arguments{
\item{file}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
mothur2 comes bundled with some example files in its \code{inst/extdata}
directory. This function make them easy to access.
}
\examples{
mothur2_example()
mothur2_example("test.fasta")
}
