% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rchime-package.R
\docType{package}
\name{rchime-package}
\alias{rchime-package}
\title{rchime: R package for chimera detection}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A package for detecting and removing chimeras.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mothur.org/rchime}
  \item Report bugs at \url{https://github.com/mothur/rchime/issues}
}

}
\author{
\strong{Maintainer}: Sarah Westcott \email{swestcot@umich.edu}

}
\keyword{internal}
