% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rchime_example.R
\name{rchime_example}
\alias{rchime_example}
\title{Get path to rchime example files}
\usage{
rchime_example(file = NULL)
}
\arguments{
\item{file}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
rchime comes bundled with some example files in its \code{inst/extdata}
directory. This function make them easy to access.
}
\examples{
rchime_example()
rchime_example("silva.gold.rds")
}
