% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rchimeReference}
\alias{rchimeReference}
\title{rchimeReference}
\usage{
rchimeReference(dataset, reference, options)
}
\arguments{
\item{dataset}{R6 'sequence_dataset' object containing sequence data}

\item{reference}{R6 'sequence_dataset' object containing sequence data}

\item{options}{list containing parameter options}
}
\description{
detects and removes chimeras from your data using a reference
dataset.
}
\seealso{
\code{\link[=rchime]{rchime()}}
}
