% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_read-sequence.R
\name{sequence}
\alias{sequence}
\title{sequence}
\description{
'sequence' is an R6 class representing a single FASTA
sequence read
}
\examples{

## ------------------------------------------------
## Method `sequence$new`
## ------------------------------------------------

fasta_read <- sequence$new("myseq", "--atgct--tt--")

## ------------------------------------------------
## Method `sequence$print`
## ------------------------------------------------

fasta_read <- sequence$new("myseq", "--atgct--tt--")
fasta_read$print()

## ------------------------------------------------
## Method `sequence$get_inLine_seq`
## ------------------------------------------------

fasta_read <- sequence$new("myseq", "--atgct--tt--")
fasta_read$get_inLine_seq()

## ------------------------------------------------
## Method `sequence$get_name`
## ------------------------------------------------

fasta_read <- sequence$new("myseq", "--atgct--tt--")
fasta_read$get_name()

## ------------------------------------------------
## Method `sequence$set_name`
## ------------------------------------------------

fasta_read <- sequence$new("myseq", "--atgct--tt--")
fasta_read$set_name("newName")

## ------------------------------------------------
## Method `sequence$set_sequence`
## ------------------------------------------------

fasta_read <- sequence$new("myseq", "--atgct--tt--")
fasta_read$set_sequence("--atgct--tt--ggcat")

## ------------------------------------------------
## Method `sequence$get_aligned`
## ------------------------------------------------

fasta_read <- sequence$new("myseq", "--atgct--tt--")
fasta_read$get_aligned()

## ------------------------------------------------
## Method `sequence$get_unaligned`
## ------------------------------------------------

fasta_read <- sequence$new("myseq", "--atgct--tt--")
fasta_read$get_unaligned()

## ------------------------------------------------
## Method `sequence$set_comment`
## ------------------------------------------------

fasta_read <- sequence$new("myseq", "--atgct--tt--")
fasta_read$set_comment("This is my comment")

## ------------------------------------------------
## Method `sequence$get_comment`
## ------------------------------------------------

fasta_read <- sequence$new("myseq", "--atgct--tt--")
fasta_read$get_comment()

## ------------------------------------------------
## Method `sequence$get_num_ambig`
## ------------------------------------------------

fasta_read <- sequence$new("myseq", "--atgct--nn--")
fasta_read$get_num_ambig()

## ------------------------------------------------
## Method `sequence$get_long_homop`
## ------------------------------------------------

fasta_read <- sequence$new("myseq", "--atgct--nn--")
fasta_read$get_long_homop()

## ------------------------------------------------
## Method `sequence$get_num_bases`
## ------------------------------------------------

fasta_read <- sequence$new("myseq", "--atgct--nn--")
fasta_read$get_num_bases()

## ------------------------------------------------
## Method `sequence$get_start`
## ------------------------------------------------

fasta_read <- sequence$new("myseq", "--atgct--nn--")
fasta_read$get_start()

## ------------------------------------------------
## Method `sequence$get_end`
## ------------------------------------------------

fasta_read <- sequence$new("myseq", "--atgct--nn--")
fasta_read$get_end()

## ------------------------------------------------
## Method `sequence$get_num_ns`
## ------------------------------------------------

fasta_read <- sequence$new("myseq", "--atgct--nn--")
fasta_read$get_num_ns()

## ------------------------------------------------
## Method `sequence$get_aligned_length`
## ------------------------------------------------

fasta_read <- sequence$new("myseq", "--atgct--nn--")
fasta_read$get_aligned_length()
}
\author{
Sarah Westcott, \email{swestcot@umich.edu}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-sequence-new}{\code{sequence$new()}}
\item \href{#method-sequence-print}{\code{sequence$print()}}
\item \href{#method-sequence-get_inLine_seq}{\code{sequence$get_inLine_seq()}}
\item \href{#method-sequence-get_name}{\code{sequence$get_name()}}
\item \href{#method-sequence-set_name}{\code{sequence$set_name()}}
\item \href{#method-sequence-set_sequence}{\code{sequence$set_sequence()}}
\item \href{#method-sequence-get_aligned}{\code{sequence$get_aligned()}}
\item \href{#method-sequence-get_unaligned}{\code{sequence$get_unaligned()}}
\item \href{#method-sequence-set_comment}{\code{sequence$set_comment()}}
\item \href{#method-sequence-get_comment}{\code{sequence$get_comment()}}
\item \href{#method-sequence-get_num_ambig}{\code{sequence$get_num_ambig()}}
\item \href{#method-sequence-get_long_homop}{\code{sequence$get_long_homop()}}
\item \href{#method-sequence-get_num_bases}{\code{sequence$get_num_bases()}}
\item \href{#method-sequence-get_start}{\code{sequence$get_start()}}
\item \href{#method-sequence-get_end}{\code{sequence$get_end()}}
\item \href{#method-sequence-get_num_ns}{\code{sequence$get_num_ns()}}
\item \href{#method-sequence-get_aligned_length}{\code{sequence$get_aligned_length()}}
\item \href{#method-sequence-clone}{\code{sequence$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence-new"></a>}}
\if{latex}{\out{\hypertarget{method-sequence-new}{}}}
\subsection{Method \code{new()}}{
Create a new sequence object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence$new(name = "", sequence = "", comment = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of sequence}

\item{\code{sequence}}{Sequence bases}

\item{\code{comment}}{Sequence comment, optional}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{sequence} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{fasta_read <- sequence$new("myseq", "--atgct--tt--")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence-print"></a>}}
\if{latex}{\out{\hypertarget{method-sequence-print}{}}}
\subsection{Method \code{print()}}{
Print a FASTA sequence
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence$print()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{fasta_read <- sequence$new("myseq", "--atgct--tt--")
fasta_read$print()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence-get_inLine_seq"></a>}}
\if{latex}{\out{\hypertarget{method-sequence-get_inLine_seq}{}}}
\subsection{Method \code{get_inLine_seq()}}{
Get string containing sequence data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence$get_inLine_seq()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{fasta_read <- sequence$new("myseq", "--atgct--tt--")
fasta_read$get_inLine_seq()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence-get_name"></a>}}
\if{latex}{\out{\hypertarget{method-sequence-get_name}{}}}
\subsection{Method \code{get_name()}}{
Get FASTA sequence name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence$get_name()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{fasta_read <- sequence$new("myseq", "--atgct--tt--")
fasta_read$get_name()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence-set_name"></a>}}
\if{latex}{\out{\hypertarget{method-sequence-set_name}{}}}
\subsection{Method \code{set_name()}}{
Set FASTA sequence name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence$set_name(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{New sequence name.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{fasta_read <- sequence$new("myseq", "--atgct--tt--")
fasta_read$set_name("newName")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence-set_sequence"></a>}}
\if{latex}{\out{\hypertarget{method-sequence-set_sequence}{}}}
\subsection{Method \code{set_sequence()}}{
Set FASTA sequence data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence$set_sequence(sequence = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sequence}}{New sequence data}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{fasta_read <- sequence$new("myseq", "--atgct--tt--")
fasta_read$set_sequence("--atgct--tt--ggcat")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence-get_aligned"></a>}}
\if{latex}{\out{\hypertarget{method-sequence-get_aligned}{}}}
\subsection{Method \code{get_aligned()}}{
Get aligned FASTA sequence data, if provided
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence$get_aligned()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{fasta_read <- sequence$new("myseq", "--atgct--tt--")
fasta_read$get_aligned()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence-get_unaligned"></a>}}
\if{latex}{\out{\hypertarget{method-sequence-get_unaligned}{}}}
\subsection{Method \code{get_unaligned()}}{
Get unaligned FASTA sequence data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence$get_unaligned()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{fasta_read <- sequence$new("myseq", "--atgct--tt--")
fasta_read$get_unaligned()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence-set_comment"></a>}}
\if{latex}{\out{\hypertarget{method-sequence-set_comment}{}}}
\subsection{Method \code{set_comment()}}{
Set FASTA sequence comment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence$set_comment(comment = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comment}}{New sequence comment}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{fasta_read <- sequence$new("myseq", "--atgct--tt--")
fasta_read$set_comment("This is my comment")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence-get_comment"></a>}}
\if{latex}{\out{\hypertarget{method-sequence-get_comment}{}}}
\subsection{Method \code{get_comment()}}{
Get FASTA sequence comment, if provided
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence$get_comment()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{fasta_read <- sequence$new("myseq", "--atgct--tt--")
fasta_read$get_comment()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence-get_num_ambig"></a>}}
\if{latex}{\out{\hypertarget{method-sequence-get_num_ambig}{}}}
\subsection{Method \code{get_num_ambig()}}{
Get number of ambiguous bases in the FASTA sequence data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence$get_num_ambig()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{fasta_read <- sequence$new("myseq", "--atgct--nn--")
fasta_read$get_num_ambig()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence-get_long_homop"></a>}}
\if{latex}{\out{\hypertarget{method-sequence-get_long_homop}{}}}
\subsection{Method \code{get_long_homop()}}{
Get length of the longest homopolymer in the FASTA sequence data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence$get_long_homop()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{fasta_read <- sequence$new("myseq", "--atgct--nn--")
fasta_read$get_long_homop()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence-get_num_bases"></a>}}
\if{latex}{\out{\hypertarget{method-sequence-get_num_bases}{}}}
\subsection{Method \code{get_num_bases()}}{
Get number of bases in the FASTA sequence data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence$get_num_bases()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{fasta_read <- sequence$new("myseq", "--atgct--nn--")
fasta_read$get_num_bases()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence-get_start"></a>}}
\if{latex}{\out{\hypertarget{method-sequence-get_start}{}}}
\subsection{Method \code{get_start()}}{
Get the alignment position of the first base in the
FASTA sequence data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence$get_start()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{fasta_read <- sequence$new("myseq", "--atgct--nn--")
fasta_read$get_start()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence-get_end"></a>}}
\if{latex}{\out{\hypertarget{method-sequence-get_end}{}}}
\subsection{Method \code{get_end()}}{
Get the alignment position of the last base in the
FASTA sequence data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence$get_end()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{fasta_read <- sequence$new("myseq", "--atgct--nn--")
fasta_read$get_end()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence-get_num_ns"></a>}}
\if{latex}{\out{\hypertarget{method-sequence-get_num_ns}{}}}
\subsection{Method \code{get_num_ns()}}{
Get the number of N's in the FASTA sequence data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence$get_num_ns()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{fasta_read <- sequence$new("myseq", "--atgct--nn--")
fasta_read$get_num_ns()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence-get_aligned_length"></a>}}
\if{latex}{\out{\hypertarget{method-sequence-get_aligned_length}{}}}
\subsection{Method \code{get_aligned_length()}}{
Get the length of the alignment of the FASTA sequence data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence$get_aligned_length()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{fasta_read <- sequence$new("myseq", "--atgct--nn--")
fasta_read$get_aligned_length()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence-clone"></a>}}
\if{latex}{\out{\hypertarget{method-sequence-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
