% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_sequence_abundance_data.R
\name{sequence_abundance_data}
\alias{sequence_abundance_data}
\title{sequence_abundance_data}
\description{
'sequence_abundance_data' is an R6 class that stores the
abundance data of sequences by sample.
}
\examples{

## ------------------------------------------------
## Method `sequence_abundance_data$print`
## ------------------------------------------------

  counts <- sequence_abundance_data$new()
  names <- c("seq1", "seq2", "seq3")
  groups <- c("sample1", "sample2", "sample1")
  counts$set_group_assignments(names, groups)
  counts$print()

## ------------------------------------------------
## Method `sequence_abundance_data$add_seqs`
## ------------------------------------------------


  dataset <- sequence_abundance_data$new()
  names <- c("seq1", "seq2", "seq3")
  dataset$add_seqs(names)


## ------------------------------------------------
## Method `sequence_abundance_data$clear`
## ------------------------------------------------

  dataset <- sequence_abundance_data$new()
  names <- c("seq1", "seq2", "seq3")
  dataset$add_seqs(names)
  dataset
  dataset$clear()
  dataset

## ------------------------------------------------
## Method `sequence_abundance_data$get_sample_table`
## ------------------------------------------------

  dataset <- sequence_abundance_data$new()
  dataset$set_group_assignments(filename =
  rchime_example("test.count_table"))

  sample_table <- dataset$get_sample_table()


## ------------------------------------------------
## Method `sequence_abundance_data$merge_seqs`
## ------------------------------------------------

  dataset <- sequence_abundance_data$new()
  names <- c("seq1", "seq2", "seq3")
  dataset$add_seqs(names)
  dataset
  dataset$merge_seqs(c("seq1", "seq2"))
  dataset

## ------------------------------------------------
## Method `sequence_abundance_data$set_abundance`
## ------------------------------------------------

  dataset <- sequence_abundance_data$new()
  names <- c("seq1", "seq2", "seq3")
  dataset$add_seqs(names)

  # set total abundance. seq1 -> 100, seq2 -> 50, seq3 -> 5
  abunds <- list(100, 50, 5)
  dataset$set_abundance("seq1", abunds[1])
  dataset$set_abundance("seq2", abunds[2])
  dataset$set_abundance("seq2", abunds[3])

  # or with group data
  dataset <- sequence_abundance_data$new()
  groups <- c("sample1", "sample1", "sample2")
  dataset$set_group_assignments(names, groups)

  # set the abundance of 'seq1' in 'sample1' to 150,
  #                      'seq1' in 'sample2' to 0
  dataset$set_abundance("seq1", list(c(150, 0)))

  # set the abundance of 'seq3' in 'sample1' to 75,
  #                      'seq3' in 'sample2' to 5
  dataset$set_abundance("seq3", list(c(75, 5)))
  dataset

## ------------------------------------------------
## Method `sequence_abundance_data$set_group_assignments`
## ------------------------------------------------

 # mothur count file
 # Representative_Sequence     total   sample2	sample3	sample4
 # seq1	1150	250	400	500
 # seq2	115	25	40	50
 # seq3	50	25	25	0
 # seq4	4	0	0	4

# as a sample table
names <- c("seq1", "seq1", "seq1",
          "seq2", "seq2", "seq2",
          "seq3", "seq3",
          "seq4")
groups <- c("sample2", "sample3", "sample4",
           "sample2", "sample3", "sample4",
           "sample2", "sample3",
           "sample4")
abundances <- c(250, 400, 500,
               25, 40, 50,
               25, 25,
               4)
abunds <- sequence_abundance_data$new()
abunds$set_group_assignments(names, groups, abundances)

# or read mothur formatted count file

abunds <- sequence_abundance_data$new()
abunds$set_group_assignments(filename =
        rchime_example("test.count_table"))

}
\author{
Sarah Westcott, \email{swestcot@umich.edu}
}
\section{Super class}{
\code{\link[rchime:sequence_dataset]{rchime::sequence_dataset}} -> \code{sequence_abundance_data}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-sequence_abundance_data-new}{\code{sequence_abundance_data$new()}}
\item \href{#method-sequence_abundance_data-print}{\code{sequence_abundance_data$print()}}
\item \href{#method-sequence_abundance_data-add_seqs}{\code{sequence_abundance_data$add_seqs()}}
\item \href{#method-sequence_abundance_data-clear}{\code{sequence_abundance_data$clear()}}
\item \href{#method-sequence_abundance_data-export}{\code{sequence_abundance_data$export()}}
\item \href{#method-sequence_abundance_data-get_abund}{\code{sequence_abundance_data$get_abund()}}
\item \href{#method-sequence_abundance_data-get_abunds}{\code{sequence_abundance_data$get_abunds()}}
\item \href{#method-sequence_abundance_data-get_groups}{\code{sequence_abundance_data$get_groups()}}
\item \href{#method-sequence_abundance_data-get_group_totals}{\code{sequence_abundance_data$get_group_totals()}}
\item \href{#method-sequence_abundance_data-get_num_groups}{\code{sequence_abundance_data$get_num_groups()}}
\item \href{#method-sequence_abundance_data-get_sample_table}{\code{sequence_abundance_data$get_sample_table()}}
\item \href{#method-sequence_abundance_data-get_total}{\code{sequence_abundance_data$get_total()}}
\item \href{#method-sequence_abundance_data-get_seqs_abunds}{\code{sequence_abundance_data$get_seqs_abunds()}}
\item \href{#method-sequence_abundance_data-has_groups}{\code{sequence_abundance_data$has_groups()}}
\item \href{#method-sequence_abundance_data-has_group}{\code{sequence_abundance_data$has_group()}}
\item \href{#method-sequence_abundance_data-merge_seqs}{\code{sequence_abundance_data$merge_seqs()}}
\item \href{#method-sequence_abundance_data-reinstate_seq}{\code{sequence_abundance_data$reinstate_seq()}}
\item \href{#method-sequence_abundance_data-remove_seq}{\code{sequence_abundance_data$remove_seq()}}
\item \href{#method-sequence_abundance_data-set_abundance}{\code{sequence_abundance_data$set_abundance()}}
\item \href{#method-sequence_abundance_data-set_group_assignments}{\code{sequence_abundance_data$set_group_assignments()}}
\item \href{#method-sequence_abundance_data-write_count_file}{\code{sequence_abundance_data$write_count_file()}}
\item \href{#method-sequence_abundance_data-clone}{\code{sequence_abundance_data$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rchime" data-topic="sequence_dataset" data-id="abort_length_mismatch"><a href='../../rchime/html/sequence_dataset.html#method-sequence_dataset-abort_length_mismatch'><code>rchime::sequence_dataset$abort_length_mismatch()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rchime" data-topic="sequence_dataset" data-id="abort_name_missing"><a href='../../rchime/html/sequence_dataset.html#method-sequence_dataset-abort_name_missing'><code>rchime::sequence_dataset$abort_name_missing()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rchime" data-topic="sequence_dataset" data-id="alert_eliminated_name"><a href='../../rchime/html/sequence_dataset.html#method-sequence_dataset-alert_eliminated_name'><code>rchime::sequence_dataset$alert_eliminated_name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rchime" data-topic="sequence_dataset" data-id="alert_missing_name"><a href='../../rchime/html/sequence_dataset.html#method-sequence_dataset-alert_missing_name'><code>rchime::sequence_dataset$alert_missing_name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rchime" data-topic="sequence_dataset" data-id="calc_ambigs"><a href='../../rchime/html/sequence_dataset.html#method-sequence_dataset-calc_ambigs'><code>rchime::sequence_dataset$calc_ambigs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rchime" data-topic="sequence_dataset" data-id="calc_end"><a href='../../rchime/html/sequence_dataset.html#method-sequence_dataset-calc_end'><code>rchime::sequence_dataset$calc_end()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rchime" data-topic="sequence_dataset" data-id="calc_longest_polymer"><a href='../../rchime/html/sequence_dataset.html#method-sequence_dataset-calc_longest_polymer'><code>rchime::sequence_dataset$calc_longest_polymer()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rchime" data-topic="sequence_dataset" data-id="calc_numbases"><a href='../../rchime/html/sequence_dataset.html#method-sequence_dataset-calc_numbases'><code>rchime::sequence_dataset$calc_numbases()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rchime" data-topic="sequence_dataset" data-id="calc_start"><a href='../../rchime/html/sequence_dataset.html#method-sequence_dataset-calc_start'><code>rchime::sequence_dataset$calc_start()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rchime" data-topic="sequence_dataset" data-id="read_fasta_file"><a href='../../rchime/html/sequence_dataset.html#method-sequence_dataset-read_fasta_file'><code>rchime::sequence_dataset$read_fasta_file()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rchime" data-topic="sequence_dataset" data-id="write_fasta_file"><a href='../../rchime/html/sequence_dataset.html#method-sequence_dataset-write_fasta_file'><code>rchime::sequence_dataset$write_fasta_file()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_abundance_data-new"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_abundance_data-new}{}}}
\subsection{Method \code{new()}}{
Create a new abundance table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_abundance_data$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{sequence_abundance_data} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_abundance_data-print"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_abundance_data-print}{}}}
\subsection{Method \code{print()}}{
View summary of count table data
'print()' outputs a summary of the sequence abundance data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_abundance_data$print()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  counts <- sequence_abundance_data$new()
  names <- c("seq1", "seq2", "seq3")
  groups <- c("sample1", "sample2", "sample1")
  counts$set_group_assignments(names, groups)
  counts$print()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_abundance_data-add_seqs"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_abundance_data-add_seqs}{}}}
\subsection{Method \code{add_seqs()}}{
Adds sequences with abundance of 1
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_abundance_data$add_seqs(names)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{names}}{a vector of sequence names}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
  dataset <- sequence_abundance_data$new()
  names <- c("seq1", "seq2", "seq3")
  dataset$add_seqs(names)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_abundance_data-clear"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_abundance_data-clear}{}}}
\subsection{Method \code{clear()}}{
Remove all sequences from dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_abundance_data$clear()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  dataset <- sequence_abundance_data$new()
  names <- c("seq1", "seq2", "seq3")
  dataset$add_seqs(names)
  dataset
  dataset$clear()
  dataset
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_abundance_data-export"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_abundance_data-export}{}}}
\subsection{Method \code{export()}}{
Get data.table containing abundance data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_abundance_data$export(names = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{names}}{vector names to export (optional)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.table
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_abundance_data-get_abund"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_abundance_data-get_abund}{}}}
\subsection{Method \code{get_abund()}}{
Get abundance for sequence in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_abundance_data$get_abund(name, group = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{String, Name of sequence}

\item{\code{group}}{(optional) Name of group to get abundance for,
if not provided the total abundance for the sequence is returned}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_abundance_data-get_abunds"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_abundance_data-get_abunds}{}}}
\subsection{Method \code{get_abunds()}}{
Get abundances for sequence by group
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_abundance_data$get_abunds(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{String, Name of sequence}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Vector of abundances, in the same order as the groups
returned by counts$get_groups().
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_abundance_data-get_groups"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_abundance_data-get_groups}{}}}
\subsection{Method \code{get_groups()}}{
Get names of groups in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_abundance_data$get_groups(name = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the sequence you want groups for, optional}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_abundance_data-get_group_totals"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_abundance_data-get_group_totals}{}}}
\subsection{Method \code{get_group_totals()}}{
Get the number of sequences represented in samples
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_abundance_data$get_group_totals(group = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group}}{String, name of sample}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of integers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_abundance_data-get_num_groups"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_abundance_data-get_num_groups}{}}}
\subsection{Method \code{get_num_groups()}}{
Get number of groups in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_abundance_data$get_num_groups()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_abundance_data-get_sample_table"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_abundance_data-get_sample_table}{}}}
\subsection{Method \code{get_sample_table()}}{
Get sample table returns data.table containing the 3 columns: name,
group and abundance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_abundance_data$get_sample_table(names = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{names, }}{vector of names to include in table (optional)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.table
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  dataset <- sequence_abundance_data$new()
  dataset$set_group_assignments(filename =
  rchime_example("test.count_table"))

  sample_table <- dataset$get_sample_table()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_abundance_data-get_total"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_abundance_data-get_total}{}}}
\subsection{Method \code{get_total()}}{
Get number of sequences in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_abundance_data$get_total()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_abundance_data-get_seqs_abunds"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_abundance_data-get_seqs_abunds}{}}}
\subsection{Method \code{get_seqs_abunds()}}{
Get abundances for sequences in dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_abundance_data$get_seqs_abunds(names, bysample = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{names}}{vector of Strings for which you want abundances for}

\item{\code{bysample}}{Boolean, if true then abundance broken down by sample,
if false, then total abundance is returned}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Vector of abundances, in the same order as the sequences
in names.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_abundance_data-has_groups"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_abundance_data-has_groups}{}}}
\subsection{Method \code{has_groups()}}{
Determine if the abundance data is divided by sample
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_abundance_data$has_groups()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Boolean
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_abundance_data-has_group"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_abundance_data-has_group}{}}}
\subsection{Method \code{has_group()}}{
Determine if a group is present in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_abundance_data$has_group(group, name = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group}}{String, Name of sample}

\item{\code{name}}{String, name of sequence. If name is provided, determine if
sequence is present in the sample (optional)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Boolean
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_abundance_data-merge_seqs"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_abundance_data-merge_seqs}{}}}
\subsection{Method \code{merge_seqs()}}{
merges identical sequences in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_abundance_data$merge_seqs(names, group = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{names}}{Vector of names to be merged}

\item{\code{group}}{String, when a group is provided only that groups counts
will be merged (optional)}

\item{\code{reason}}{String, reason for merging sequence}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  dataset <- sequence_abundance_data$new()
  names <- c("seq1", "seq2", "seq3")
  dataset$add_seqs(names)
  dataset
  dataset$merge_seqs(c("seq1", "seq2"))
  dataset
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_abundance_data-reinstate_seq"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_abundance_data-reinstate_seq}{}}}
\subsection{Method \code{reinstate_seq()}}{
Reinstates sequence in active table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_abundance_data$reinstate_seq(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{name of sequence to reinstate}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_abundance_data-remove_seq"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_abundance_data-remove_seq}{}}}
\subsection{Method \code{remove_seq()}}{
Remove sequences from dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_abundance_data$remove_seq(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{String, name of sequence to remove}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_abundance_data-set_abundance"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_abundance_data-set_abundance}{}}}
\subsection{Method \code{set_abundance()}}{
Set abundance data for sequence
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_abundance_data$set_abundance(name, abundance)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{a vector of sequence names}

\item{\code{abundance}}{a list of abundance assignments}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  dataset <- sequence_abundance_data$new()
  names <- c("seq1", "seq2", "seq3")
  dataset$add_seqs(names)

  # set total abundance. seq1 -> 100, seq2 -> 50, seq3 -> 5
  abunds <- list(100, 50, 5)
  dataset$set_abundance("seq1", abunds[1])
  dataset$set_abundance("seq2", abunds[2])
  dataset$set_abundance("seq2", abunds[3])

  # or with group data
  dataset <- sequence_abundance_data$new()
  groups <- c("sample1", "sample1", "sample2")
  dataset$set_group_assignments(names, groups)

  # set the abundance of 'seq1' in 'sample1' to 150,
  #                      'seq1' in 'sample2' to 0
  dataset$set_abundance("seq1", list(c(150, 0)))

  # set the abundance of 'seq3' in 'sample1' to 75,
  #                      'seq3' in 'sample2' to 5
  dataset$set_abundance("seq3", list(c(75, 5)))
  dataset
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_abundance_data-set_group_assignments"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_abundance_data-set_group_assignments}{}}}
\subsection{Method \code{set_group_assignments()}}{
Add group assignment data, clears out old sample data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_abundance_data$set_group_assignments(
  names = NULL,
  groups = NULL,
  abundances = NULL,
  filename = NULL,
  path = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{names}}{a vector of sequence names}

\item{\code{groups}}{a vector of group assignments}

\item{\code{abundances}}{a vector of sample abundances}

\item{\code{filename}}{String, Name of mothur formatted count file}

\item{\code{path}}{String, Path to mothur formatted count file}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{ # mothur count file
 # Representative_Sequence     total   sample2	sample3	sample4
 # seq1	1150	250	400	500
 # seq2	115	25	40	50
 # seq3	50	25	25	0
 # seq4	4	0	0	4

# as a sample table
names <- c("seq1", "seq1", "seq1",
          "seq2", "seq2", "seq2",
          "seq3", "seq3",
          "seq4")
groups <- c("sample2", "sample3", "sample4",
           "sample2", "sample3", "sample4",
           "sample2", "sample3",
           "sample4")
abundances <- c(250, 400, 500,
               25, 40, 50,
               25, 25,
               4)
abunds <- sequence_abundance_data$new()
abunds$set_group_assignments(names, groups, abundances)

# or read mothur formatted count file

abunds <- sequence_abundance_data$new()
abunds$set_group_assignments(filename =
        rchime_example("test.count_table"))

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_abundance_data-write_count_file"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_abundance_data-write_count_file}{}}}
\subsection{Method \code{write_count_file()}}{
Write count table data to mothur formatted count file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_abundance_data$write_count_file(filename, path = NULL, fasta_names)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{String, name of file you want to write to}

\item{\code{path}}{path to count file (optional)}

\item{\code{fasta_names}}{vector of strings containing names to write}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_abundance_data-clone"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_abundance_data-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_abundance_data$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
