% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_sequence_data_vector.R
\name{sequence_data_vector}
\alias{sequence_data_vector}
\title{sequence_data_vector}
\description{
'sequence_data_vector' is an R6 class that represents sequence
FASTA and abundance data. It inherits from 'sequence_dataset'.
}
\examples{

## ------------------------------------------------
## Method `sequence_data_vector$new`
## ------------------------------------------------

  dataset <- sequence_data_vector$new(filename =
    rchime_example("test.fasta"))


## ------------------------------------------------
## Method `sequence_data_vector$print`
## ------------------------------------------------

  dataset <- sequence_data_vector$new(filename =
  rchime_example("test.fasta"))
  dataset$print()

## ------------------------------------------------
## Method `sequence_data_vector$print_bad_accnos`
## ------------------------------------------------

  dataset <- sequence_data_vector$new(filename =
    rchime_example("test.fasta"))

  # flag 10 'bad' sequences
  names <- dataset$get_names()[c(1:10)]
  trash_codes <- rep("first_ten", 10)

  dataset$remove_seqs(names, trash_codes)

  dataset$print_bad_accnos()

## ------------------------------------------------
## Method `sequence_data_vector$add_seqs`
## ------------------------------------------------


  dataset <- sequence_data_vector$new()
  names <- c("seq1", "seq2", "seq3")
  sequences <- c("ATGGGCT", "..TG--ACCGT..", "..GGuatgc..")
  comments <- c("", "myComment", "")
  dataset$add_seqs(names, sequences, comments)


## ------------------------------------------------
## Method `sequence_data_vector$clear`
## ------------------------------------------------

  dataset <- sequence_data_vector$new(filename =
  rchime_example("test.fasta"))
  dataset$clear()

## ------------------------------------------------
## Method `sequence_data_vector$get_abund`
## ------------------------------------------------

  dataset <- sequence_data_vector$new(filename =
  rchime_example("test.fasta"))
  dataset <- dataset$set_group_assignments(filename =
  rchime_example("test.count_table"))
  dataset$get_abund(name =
   "M00967_43_000000000-A3JHG_1_1111_8697_7063")

## ------------------------------------------------
## Method `sequence_data_vector$get_abunds`
## ------------------------------------------------

  dataset <- sequence_data_vector$new(filename =
  rchime_example("test.fasta"))
  dataset <- dataset$set_group_assignments(filename =
  rchime_example("test.count_table"))
  dataset$get_abunds(name =
   "M00967_43_000000000-A3JHG_1_1111_8697_7063")

## ------------------------------------------------
## Method `sequence_data_vector$get_accnos_report`
## ------------------------------------------------

 dataset <- sequence_data_vector$new(filename =
 rchime_example("test.fasta"))
 some_seqs_to_remove <- c("M00967_43_000000000-A3JHG_1_2113_21675_17687",
 "M00967_43_000000000-A3JHG_1_1114_17672_13068")
 reasons <- c("example_removal", "example_removal")
 dataset$remove_seqs(some_seqs_to_remove, reasons)
 df <- dataset$get_accnos_report()

## ------------------------------------------------
## Method `sequence_data_vector$get_align_report`
## ------------------------------------------------

  dataset <- sequence_data_vector$new(filename =
    rchime_example("test.fasta"))
  reference <- sequence_data_vector$new(filename =
    rchime_example("reference.fasta"))

  # results$align_report contains complete align report
  # results$accnos will contain the names of any sequences that were
  #   flipped

  # results <- align_seqs(dataset, reference)

  # report contains partial align report used for screening:
  # search_score, sim_score and longest_insert
  report <- dataset$get_align_report()


## ------------------------------------------------
## Method `sequence_data_vector$get_contigs_report`
## ------------------------------------------------

  # You can use the mothur2 package to assemble your
  # paired reads and create a dataset with contigs data

  # files <- make_file(input_dir = rchime_example(), type = "gz")
  # contigs <- make_contigs(files)

  # df <- contigs$dataset$get_contigs_report()


## ------------------------------------------------
## Method `sequence_data_vector$get_count_table`
## ------------------------------------------------

  dataset <- sequence_data_vector$new(filename =
  rchime_example("test.fasta"))
  dataset$set_group_assignments(filename =
  rchime_example("test.count_table"))
  count_table <- dataset$get_count_table()

## ------------------------------------------------
## Method `sequence_data_vector$get_names`
## ------------------------------------------------

  dataset <- sequence_data_vector$new(
  filename = rchime_example('test.fasta'))

  dataset <- dataset$set_group_assignments(filename =
  rchime_example("test.count_table"))

  # get names of sequences in your dataset
  names <- dataset$get_names()

  # get names of sequences in sample 'F3D0'
  names <- dataset$get_names(group = 'F3D0')


## ------------------------------------------------
## Method `sequence_data_vector$get_seqs_table`
## ------------------------------------------------

  dataset <- sequence_data_vector$new(filename =
  rchime_example("test.fasta"))
  seqs <- dataset$get_seqs_table()

## ------------------------------------------------
## Method `sequence_data_vector$get_seqs`
## ------------------------------------------------

  dataset <- sequence_data_vector$new(filename =
  rchime_example("test.fasta"))

  # to get all sequences in dataset
  seqs <- dataset$get_seqs()

  # to get sequences from sample 'F3D0'
  seqs <- dataset$get_seqs(group = 'F3D0')


## ------------------------------------------------
## Method `sequence_data_vector$get_seqs_abunds`
## ------------------------------------------------

  dataset <- sequence_data_vector$new(
  filename = rchime_example('test.fasta'))

  dataset <- dataset$set_group_assignments(filename =
  rchime_example("test.count_table"))

  # get abundances of all sequences in your dataset
  abunds <- dataset$get_seqs_abunds()

  # get abundances of sequences divided by sample
  abunds <- dataset$get_seqs_abunds(bysample = TRUE)

## ------------------------------------------------
## Method `sequence_data_vector$get_summary_report`
## ------------------------------------------------

  dataset <- sequence_data_vector$new(filename =
  rchime_example("test.fasta"))
  dataset <- dataset$set_group_assignments(filename =
  rchime_example("test.count_table"))
  df <- dataset$get_summary_report()

## ------------------------------------------------
## Method `sequence_data_vector$merge_seqs`
## ------------------------------------------------

  dataset <- sequence_data_vector$new(filename =
  rchime_example("test.fasta"))
  dataset$set_group_assignments(filename =
  rchime_example("test.count_table"))

  # Number of Unique seqs: 1000
  # Total number of seqs: 1000

  seqs_to_merge <- list(c("M00967_43_000000000-A3JHG_1_1112_15471_15261",
  "M00967_43_000000000-A3JHG_1_2114_21028_11825",
  "M00967_43_000000000-A3JHG_1_2106_5820_9300"),
  c("M00967_43_000000000-A3JHG_1_1111_26505_10426",
  "M00967_43_000000000-A3JHG_1_1111_23613_8947"))

  dataset$merge_seqs(seqs_to_merge)

  # Number of Unique seqs: 997
  # Total number of seqs: 1000

## ------------------------------------------------
## Method `sequence_data_vector$reinstate_seqs`
## ------------------------------------------------

  dataset <- sequence_data_vector$new(filename =
  rchime_example("test.fasta"))
  dataset$set_group_assignments(filename =
  rchime_example("test.count_table"))

  # flag 10 sequences for removal
  names <- dataset$get_names()[c(1:10)]
  trash_codes <- rep("first_ten", 10)

  dataset$remove_seqs(names, trash_codes)

  # reinclude sequences you removed
  dataset$reinstate_seqs(trash_codes = c("first_ten"))

## ------------------------------------------------
## Method `sequence_data_vector$remove_seqs`
## ------------------------------------------------

  dataset <- sequence_data_vector$new(filename =
                    rchime_example("test.fasta"))
  dataset$set_group_assignments(filename =
                    rchime_example("test.count_table"))
  dataset

# remove first 10 sequences from the dataset

  names <- dataset$get_names()[c(1:10)]
  trash_codes <- rep("first_ten", 10)

  dataset$remove_seqs(names, trash_codes)
  dataset


## ------------------------------------------------
## Method `sequence_data_vector$set_group_assignments`
## ------------------------------------------------


 # mothur count file
 # Representative_Sequence     total   sample2	sample3	sample4
 # seq1	1150	250	400	500
 # seq2	115	25	40	50
 # seq3	50	25	25	0
 # seq4	4	0	0	4

# inputted as a sample table
names <- c("seq1", "seq1", "seq1",
          "seq2", "seq2", "seq2",
          "seq3", "seq3",
          "seq4")
groups <- c("sample2", "sample3", "sample4",
           "sample2", "sample3", "sample4",
           "sample2", "sample3",
           "sample4")
abundances <- c(250, 400, 500,
               25, 40, 50,
               25, 25,
               4)

dataset <- sequence_data_vector$new()
unique_names <- unique(names)
sequences <- c("ATGGGCT", "..TG--ACCGT..", "..GGuatgc..", "..GGTAC-T..")
dataset$add_seqs(unique_names, sequences)
dataset$set_group_assignments(names, groups, abundances)

# or read fasta file and mothur formatted count file

dataset <- sequence_data_vector$new(filename =
        rchime_example("test.fasta"))
dataset$set_group_assignments(filename =
        rchime_example("test.count_table"))


## ------------------------------------------------
## Method `sequence_data_vector$set_abundances`
## ------------------------------------------------

  dataset <- sequence_data_vector$new()
  names <- c("seq1", "seq2", "seq3")
  sequences <- c("ATGGGCT", "..TG--ACCGT..", "..GGuatgc..")

  dataset$add_seqs(names, sequences)

  # set total abundance. seq1 -> 100, seq2 -> 50, seq3 -> 5
  abunds <- list(100, 50, 5)
  dataset$set_abundances(names, abunds)

  # add group data
  groups <- c("sample1", "sample1", "sample2")
  dataset$set_group_assignments(names, groups)

  # set the abundance of 'seq1' in 'sample1' to 150
  dataset$set_abundances(c("seq1"), list(c(150, 0)))

  # set the abundance of 'seq1' in 'sample1' to 150,
  #                      'seq1' in 'sample2' to 0 and
  # set the abundance of 'seq3' in 'sample1' to 75,
  #                      'seq3' in 'sample2' to 5
  dataset$set_abundances(c("seq1", "seq3"), list(c(150, 0), c(75, 5)))


## ------------------------------------------------
## Method `sequence_data_vector$set_seqs`
## ------------------------------------------------

  dataset <- sequence_data_vector$new()
  names <- c("seq1", "seq2", "seq3")
  sequences <- c("ATGGGCT", "..TG--ACCGT..", "..GGuatgc..")
  dataset$add_seqs(names, sequences)

  # change sequence string for seq2 from ..TG--ACCGT.. to ...ATG-G-GCT..
  dataset$set_seqs(c("seq2"), c("...ATG-G-GCT.."))

}
\author{
Sarah Westcott, \email{swestcot@umich.edu}
}
\section{Super class}{
\code{\link[rchime:sequence_dataset]{rchime::sequence_dataset}} -> \code{sequence_data_vector}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-sequence_data_vector-new}{\code{sequence_data_vector$new()}}
\item \href{#method-sequence_data_vector-print}{\code{sequence_data_vector$print()}}
\item \href{#method-sequence_data_vector-print_bad_accnos}{\code{sequence_data_vector$print_bad_accnos()}}
\item \href{#method-sequence_data_vector-add_seqs}{\code{sequence_data_vector$add_seqs()}}
\item \href{#method-sequence_data_vector-clear}{\code{sequence_data_vector$clear()}}
\item \href{#method-sequence_data_vector-get_abund}{\code{sequence_data_vector$get_abund()}}
\item \href{#method-sequence_data_vector-get_abunds}{\code{sequence_data_vector$get_abunds()}}
\item \href{#method-sequence_data_vector-get_accnos_report}{\code{sequence_data_vector$get_accnos_report()}}
\item \href{#method-sequence_data_vector-get_align_report}{\code{sequence_data_vector$get_align_report()}}
\item \href{#method-sequence_data_vector-get_contigs_report}{\code{sequence_data_vector$get_contigs_report()}}
\item \href{#method-sequence_data_vector-get_count_table}{\code{sequence_data_vector$get_count_table()}}
\item \href{#method-sequence_data_vector-get_groups}{\code{sequence_data_vector$get_groups()}}
\item \href{#method-sequence_data_vector-get_group_totals}{\code{sequence_data_vector$get_group_totals()}}
\item \href{#method-sequence_data_vector-get_names}{\code{sequence_data_vector$get_names()}}
\item \href{#method-sequence_data_vector-get_num_groups}{\code{sequence_data_vector$get_num_groups()}}
\item \href{#method-sequence_data_vector-get_num_seqs}{\code{sequence_data_vector$get_num_seqs()}}
\item \href{#method-sequence_data_vector-get_num_unique_seqs}{\code{sequence_data_vector$get_num_unique_seqs()}}
\item \href{#method-sequence_data_vector-get_seqs_table}{\code{sequence_data_vector$get_seqs_table()}}
\item \href{#method-sequence_data_vector-get_seqs}{\code{sequence_data_vector$get_seqs()}}
\item \href{#method-sequence_data_vector-get_seqs_abunds}{\code{sequence_data_vector$get_seqs_abunds()}}
\item \href{#method-sequence_data_vector-get_summary_report}{\code{sequence_data_vector$get_summary_report()}}
\item \href{#method-sequence_data_vector-has_align_report}{\code{sequence_data_vector$has_align_report()}}
\item \href{#method-sequence_data_vector-has_contigs_report}{\code{sequence_data_vector$has_contigs_report()}}
\item \href{#method-sequence_data_vector-has_group}{\code{sequence_data_vector$has_group()}}
\item \href{#method-sequence_data_vector-in_table}{\code{sequence_data_vector$in_table()}}
\item \href{#method-sequence_data_vector-is_aligned}{\code{sequence_data_vector$is_aligned()}}
\item \href{#method-sequence_data_vector-merge_seqs}{\code{sequence_data_vector$merge_seqs()}}
\item \href{#method-sequence_data_vector-reinstate_seqs}{\code{sequence_data_vector$reinstate_seqs()}}
\item \href{#method-sequence_data_vector-remove_seqs}{\code{sequence_data_vector$remove_seqs()}}
\item \href{#method-sequence_data_vector-set_group_assignments}{\code{sequence_data_vector$set_group_assignments()}}
\item \href{#method-sequence_data_vector-set_abundances}{\code{sequence_data_vector$set_abundances()}}
\item \href{#method-sequence_data_vector-set_seqs}{\code{sequence_data_vector$set_seqs()}}
\item \href{#method-sequence_data_vector-write_count_file}{\code{sequence_data_vector$write_count_file()}}
\item \href{#method-sequence_data_vector-clone}{\code{sequence_data_vector$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rchime" data-topic="sequence_dataset" data-id="abort_length_mismatch"><a href='../../rchime/html/sequence_dataset.html#method-sequence_dataset-abort_length_mismatch'><code>rchime::sequence_dataset$abort_length_mismatch()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rchime" data-topic="sequence_dataset" data-id="abort_name_missing"><a href='../../rchime/html/sequence_dataset.html#method-sequence_dataset-abort_name_missing'><code>rchime::sequence_dataset$abort_name_missing()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rchime" data-topic="sequence_dataset" data-id="alert_eliminated_name"><a href='../../rchime/html/sequence_dataset.html#method-sequence_dataset-alert_eliminated_name'><code>rchime::sequence_dataset$alert_eliminated_name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rchime" data-topic="sequence_dataset" data-id="alert_missing_name"><a href='../../rchime/html/sequence_dataset.html#method-sequence_dataset-alert_missing_name'><code>rchime::sequence_dataset$alert_missing_name()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rchime" data-topic="sequence_dataset" data-id="calc_ambigs"><a href='../../rchime/html/sequence_dataset.html#method-sequence_dataset-calc_ambigs'><code>rchime::sequence_dataset$calc_ambigs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rchime" data-topic="sequence_dataset" data-id="calc_end"><a href='../../rchime/html/sequence_dataset.html#method-sequence_dataset-calc_end'><code>rchime::sequence_dataset$calc_end()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rchime" data-topic="sequence_dataset" data-id="calc_longest_polymer"><a href='../../rchime/html/sequence_dataset.html#method-sequence_dataset-calc_longest_polymer'><code>rchime::sequence_dataset$calc_longest_polymer()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rchime" data-topic="sequence_dataset" data-id="calc_numbases"><a href='../../rchime/html/sequence_dataset.html#method-sequence_dataset-calc_numbases'><code>rchime::sequence_dataset$calc_numbases()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rchime" data-topic="sequence_dataset" data-id="calc_start"><a href='../../rchime/html/sequence_dataset.html#method-sequence_dataset-calc_start'><code>rchime::sequence_dataset$calc_start()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rchime" data-topic="sequence_dataset" data-id="read_fasta_file"><a href='../../rchime/html/sequence_dataset.html#method-sequence_dataset-read_fasta_file'><code>rchime::sequence_dataset$read_fasta_file()</code></a></span></li>
<li><span class="pkg-link" data-pkg="rchime" data-topic="sequence_dataset" data-id="write_fasta_file"><a href='../../rchime/html/sequence_dataset.html#method-sequence_dataset-write_fasta_file'><code>rchime::sequence_dataset$write_fasta_file()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data_vector-new"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data_vector-new}{}}}
\subsection{Method \code{new()}}{
Create a new FASTA sequence dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data_vector$new(filename = NULL, path = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{Name of file containing FASTA reads}

\item{\code{path}}{Path to file containing FASTA reads}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{sequence_data_vector} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  dataset <- sequence_data_vector$new(filename =
    rchime_example("test.fasta"))

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data_vector-print"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data_vector-print}{}}}
\subsection{Method \code{print()}}{
Get summary of sequences data - (summary_seqs output)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data_vector$print()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  dataset <- sequence_data_vector$new(filename =
  rchime_example("test.fasta"))
  dataset$print()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data_vector-print_bad_accnos"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data_vector-print_bad_accnos}{}}}
\subsection{Method \code{print_bad_accnos()}}{
View summary of removed sequences
'print_bad_accnos()' outputs a summary of the reasons sequences were
removed from the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data_vector$print_bad_accnos()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  dataset <- sequence_data_vector$new(filename =
    rchime_example("test.fasta"))

  # flag 10 'bad' sequences
  names <- dataset$get_names()[c(1:10)]
  trash_codes <- rep("first_ten", 10)

  dataset$remove_seqs(names, trash_codes)

  dataset$print_bad_accnos()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data_vector-add_seqs"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data_vector-add_seqs}{}}}
\subsection{Method \code{add_seqs()}}{
Add new sequence data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data_vector$add_seqs(
  names,
  sequences,
  comments = c(),
  olengths = c(),
  ostarts = c(),
  oends = c(),
  mismatches = c(),
  ee = c()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{names}}{a vector of sequence names}

\item{\code{sequences}}{a vector of sequence data}

\item{\code{comments}}{a vector of sequence comments, (optional)}

\item{\code{olengths}}{a vector containing sequences overlap lengths,
(optional)}

\item{\code{ostarts}}{a vector containing sequences overlap starts, optional}

\item{\code{oends}}{a vector containing sequences overlap ends, optional}

\item{\code{mismatches}}{a vector containing the number of sequence
mismatches, (optional)}

\item{\code{ee}}{a vector containing the sequences expected errors, optional}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
  dataset <- sequence_data_vector$new()
  names <- c("seq1", "seq2", "seq3")
  sequences <- c("ATGGGCT", "..TG--ACCGT..", "..GGuatgc..")
  comments <- c("", "myComment", "")
  dataset$add_seqs(names, sequences, comments)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data_vector-clear"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data_vector-clear}{}}}
\subsection{Method \code{clear()}}{
Remove all sequences from dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data_vector$clear()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  dataset <- sequence_data_vector$new(filename =
  rchime_example("test.fasta"))
  dataset$clear()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data_vector-get_abund"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data_vector-get_abund}{}}}
\subsection{Method \code{get_abund()}}{
Get abundance for sequence in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data_vector$get_abund(name, group = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{String, Name of sequence}

\item{\code{group}}{(optional) Name of group to get abundance for,
if not provided the total abundance for the sequence is returned}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An integer
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  dataset <- sequence_data_vector$new(filename =
  rchime_example("test.fasta"))
  dataset <- dataset$set_group_assignments(filename =
  rchime_example("test.count_table"))
  dataset$get_abund(name =
   "M00967_43_000000000-A3JHG_1_1111_8697_7063")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data_vector-get_abunds"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data_vector-get_abunds}{}}}
\subsection{Method \code{get_abunds()}}{
Get abundances for sequence by group
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data_vector$get_abunds(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{String, Name of sequence}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Vector of abundances, in the same order as the groups
returned by dataset$get_groups().
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  dataset <- sequence_data_vector$new(filename =
  rchime_example("test.fasta"))
  dataset <- dataset$set_group_assignments(filename =
  rchime_example("test.count_table"))
  dataset$get_abunds(name =
   "M00967_43_000000000-A3JHG_1_1111_8697_7063")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data_vector-get_accnos_report"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data_vector-get_accnos_report}{}}}
\subsection{Method \code{get_accnos_report()}}{
Get data.table containing a report of sequences removed
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data_vector$get_accnos_report()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{ dataset <- sequence_data_vector$new(filename =
 rchime_example("test.fasta"))
 some_seqs_to_remove <- c("M00967_43_000000000-A3JHG_1_2113_21675_17687",
 "M00967_43_000000000-A3JHG_1_1114_17672_13068")
 reasons <- c("example_removal", "example_removal")
 dataset$remove_seqs(some_seqs_to_remove, reasons)
 df <- dataset$get_accnos_report()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data_vector-get_align_report"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data_vector-get_align_report}{}}}
\subsection{Method \code{get_align_report()}}{
Get data.table containing align report data used for screening
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data_vector$get_align_report()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  dataset <- sequence_data_vector$new(filename =
    rchime_example("test.fasta"))
  reference <- sequence_data_vector$new(filename =
    rchime_example("reference.fasta"))

  # results$align_report contains complete align report
  # results$accnos will contain the names of any sequences that were
  #   flipped

  # results <- align_seqs(dataset, reference)

  # report contains partial align report used for screening:
  # search_score, sim_score and longest_insert
  report <- dataset$get_align_report()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data_vector-get_contigs_report"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data_vector-get_contigs_report}{}}}
\subsection{Method \code{get_contigs_report()}}{
Get data.table containing contigs report data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data_vector$get_contigs_report()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  # You can use the mothur2 package to assemble your
  # paired reads and create a dataset with contigs data

  # files <- make_file(input_dir = rchime_example(), type = "gz")
  # contigs <- make_contigs(files)

  # df <- contigs$dataset$get_contigs_report()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data_vector-get_count_table"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data_vector-get_count_table}{}}}
\subsection{Method \code{get_count_table()}}{
Get count table data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data_vector$get_count_table()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a data.table containing the count table data
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  dataset <- sequence_data_vector$new(filename =
  rchime_example("test.fasta"))
  dataset$set_group_assignments(filename =
  rchime_example("test.count_table"))
  count_table <- dataset$get_count_table()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data_vector-get_groups"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data_vector-get_groups}{}}}
\subsection{Method \code{get_groups()}}{
Get names of groups in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data_vector$get_groups(name = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the sequence you want groups for, optional}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data_vector-get_group_totals"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data_vector-get_group_totals}{}}}
\subsection{Method \code{get_group_totals()}}{
Get the number of sequences represented in samples
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data_vector$get_group_totals(group = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group}}{String, name of sample}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of integers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data_vector-get_names"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data_vector-get_names}{}}}
\subsection{Method \code{get_names()}}{
Get names of sequences in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data_vector$get_names(group = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group}}{String, name of sample}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  dataset <- sequence_data_vector$new(
  filename = rchime_example('test.fasta'))

  dataset <- dataset$set_group_assignments(filename =
  rchime_example("test.count_table"))

  # get names of sequences in your dataset
  names <- dataset$get_names()

  # get names of sequences in sample 'F3D0'
  names <- dataset$get_names(group = 'F3D0')

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data_vector-get_num_groups"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data_vector-get_num_groups}{}}}
\subsection{Method \code{get_num_groups()}}{
Get number of groups in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data_vector$get_num_groups()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data_vector-get_num_seqs"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data_vector-get_num_seqs}{}}}
\subsection{Method \code{get_num_seqs()}}{
Get the number of sequences in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data_vector$get_num_seqs(group = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group}}{The name of the group you want number of sequences for,
optional}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data_vector-get_num_unique_seqs"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data_vector-get_num_unique_seqs}{}}}
\subsection{Method \code{get_num_unique_seqs()}}{
Get number of unique sequences in dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data_vector$get_num_unique_seqs()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data_vector-get_seqs_table"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data_vector-get_seqs_table}{}}}
\subsection{Method \code{get_seqs_table()}}{
Get data.table containing FASTA data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data_vector$get_seqs_table()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data.table
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  dataset <- sequence_data_vector$new(filename =
  rchime_example("test.fasta"))
  seqs <- dataset$get_seqs_table()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data_vector-get_seqs"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data_vector-get_seqs}{}}}
\subsection{Method \code{get_seqs()}}{
Get vector containing FASTA nucleotide strings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data_vector$get_seqs(group = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group}}{String, name of sample}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
data.table
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  dataset <- sequence_data_vector$new(filename =
  rchime_example("test.fasta"))

  # to get all sequences in dataset
  seqs <- dataset$get_seqs()

  # to get sequences from sample 'F3D0'
  seqs <- dataset$get_seqs(group = 'F3D0')

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data_vector-get_seqs_abunds"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data_vector-get_seqs_abunds}{}}}
\subsection{Method \code{get_seqs_abunds()}}{
Get abundances for sequences in dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data_vector$get_seqs_abunds(bysample = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bysample}}{Boolean, if true then abundance broken down by sample,
if false, then total abundance is returned}

\item{\code{names}}{vector of Strings for which you want abundances for}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Vector of abundances, in the same order as the sequences
returned by dataset$get_names().
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  dataset <- sequence_data_vector$new(
  filename = rchime_example('test.fasta'))

  dataset <- dataset$set_group_assignments(filename =
  rchime_example("test.count_table"))

  # get abundances of all sequences in your dataset
  abunds <- dataset$get_seqs_abunds()

  # get abundances of sequences divided by sample
  abunds <- dataset$get_seqs_abunds(bysample = TRUE)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data_vector-get_summary_report"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data_vector-get_summary_report}{}}}
\subsection{Method \code{get_summary_report()}}{
Get data.table containing summary report data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data_vector$get_summary_report()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  dataset <- sequence_data_vector$new(filename =
  rchime_example("test.fasta"))
  dataset <- dataset$set_group_assignments(filename =
  rchime_example("test.count_table"))
  df <- dataset$get_summary_report()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data_vector-has_align_report"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data_vector-has_align_report}{}}}
\subsection{Method \code{has_align_report()}}{
Determine if the dataset contains align report data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data_vector$has_align_report()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Boolean
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data_vector-has_contigs_report"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data_vector-has_contigs_report}{}}}
\subsection{Method \code{has_contigs_report()}}{
Determine if the dataset contains contigs report data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data_vector$has_contigs_report()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Boolean
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data_vector-has_group"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data_vector-has_group}{}}}
\subsection{Method \code{has_group()}}{
Determine if a group is present in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data_vector$has_group(group)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group}}{String, Name of sample}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Boolean
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data_vector-in_table"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data_vector-in_table}{}}}
\subsection{Method \code{in_table()}}{
Determine if a sequence is present in the dataset and its index
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data_vector$in_table(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{String, Name of sequence}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
list, list(include = boolean, index = row in table)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data_vector-is_aligned"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data_vector-is_aligned}{}}}
\subsection{Method \code{is_aligned()}}{
Determine if the dataset is aligned and its alignment length
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data_vector$is_aligned()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list(aligned = (TRUE OR FALSE),
length = (alignment length or NA))
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data_vector-merge_seqs"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data_vector-merge_seqs}{}}}
\subsection{Method \code{merge_seqs()}}{
merges identical sequences in the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data_vector$merge_seqs(seqs_to_merge, reason = "merged", group = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seqs_to_merge}}{List of vectors, each vector will be merged}

\item{\code{reason}}{String, reason for merging sequence}

\item{\code{group}}{String, when a group is provided only that groups counts
will be merged (optional)}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  dataset <- sequence_data_vector$new(filename =
  rchime_example("test.fasta"))
  dataset$set_group_assignments(filename =
  rchime_example("test.count_table"))

  # Number of Unique seqs: 1000
  # Total number of seqs: 1000

  seqs_to_merge <- list(c("M00967_43_000000000-A3JHG_1_1112_15471_15261",
  "M00967_43_000000000-A3JHG_1_2114_21028_11825",
  "M00967_43_000000000-A3JHG_1_2106_5820_9300"),
  c("M00967_43_000000000-A3JHG_1_1111_26505_10426",
  "M00967_43_000000000-A3JHG_1_1111_23613_8947"))

  dataset$merge_seqs(seqs_to_merge)

  # Number of Unique seqs: 997
  # Total number of seqs: 1000
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data_vector-reinstate_seqs"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data_vector-reinstate_seqs}{}}}
\subsection{Method \code{reinstate_seqs()}}{
Reinstates sequences removed for the reasons in trash_codes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data_vector$reinstate_seqs(trash_codes)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{trash_codes}}{vector containing reasons for sequences removal}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  dataset <- sequence_data_vector$new(filename =
  rchime_example("test.fasta"))
  dataset$set_group_assignments(filename =
  rchime_example("test.count_table"))

  # flag 10 sequences for removal
  names <- dataset$get_names()[c(1:10)]
  trash_codes <- rep("first_ten", 10)

  dataset$remove_seqs(names, trash_codes)

  # reinclude sequences you removed
  dataset$reinstate_seqs(trash_codes = c("first_ten"))
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data_vector-remove_seqs"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data_vector-remove_seqs}{}}}
\subsection{Method \code{remove_seqs()}}{
Remove sequences from dataset for cause
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data_vector$remove_seqs(names, trash_codes)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{names}}{vector containing names of sequences to remove}

\item{\code{trash_codes}}{vector containing reasons for sequences removal}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  dataset <- sequence_data_vector$new(filename =
                    rchime_example("test.fasta"))
  dataset$set_group_assignments(filename =
                    rchime_example("test.count_table"))
  dataset

# remove first 10 sequences from the dataset

  names <- dataset$get_names()[c(1:10)]
  trash_codes <- rep("first_ten", 10)

  dataset$remove_seqs(names, trash_codes)
  dataset

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data_vector-set_group_assignments"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data_vector-set_group_assignments}{}}}
\subsection{Method \code{set_group_assignments()}}{
Add group assignment data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data_vector$set_group_assignments(
  names = NULL,
  groups = NULL,
  abundances = NULL,
  filename = NULL,
  path = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{names}}{a vector of sequence names}

\item{\code{groups}}{a vector of group assignments}

\item{\code{abundances}}{a vector of sample abundances}

\item{\code{filename}}{String, Name of mothur formatted count file}

\item{\code{path}}{String, Path to mothur formatted count file}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{
 # mothur count file
 # Representative_Sequence     total   sample2	sample3	sample4
 # seq1	1150	250	400	500
 # seq2	115	25	40	50
 # seq3	50	25	25	0
 # seq4	4	0	0	4

# inputted as a sample table
names <- c("seq1", "seq1", "seq1",
          "seq2", "seq2", "seq2",
          "seq3", "seq3",
          "seq4")
groups <- c("sample2", "sample3", "sample4",
           "sample2", "sample3", "sample4",
           "sample2", "sample3",
           "sample4")
abundances <- c(250, 400, 500,
               25, 40, 50,
               25, 25,
               4)

dataset <- sequence_data_vector$new()
unique_names <- unique(names)
sequences <- c("ATGGGCT", "..TG--ACCGT..", "..GGuatgc..", "..GGTAC-T..")
dataset$add_seqs(unique_names, sequences)
dataset$set_group_assignments(names, groups, abundances)

# or read fasta file and mothur formatted count file

dataset <- sequence_data_vector$new(filename =
        rchime_example("test.fasta"))
dataset$set_group_assignments(filename =
        rchime_example("test.count_table"))

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data_vector-set_abundances"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data_vector-set_abundances}{}}}
\subsection{Method \code{set_abundances()}}{
Set abundance data for sequences
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data_vector$set_abundances(
  names = NULL,
  abunds = NULL,
  reason = "merged"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{names}}{a vector of sequence names}

\item{\code{abunds}}{a list of abundance assignments}

\item{\code{reason, }}{String (optional) reason for setting sequences to
zero abundance}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  dataset <- sequence_data_vector$new()
  names <- c("seq1", "seq2", "seq3")
  sequences <- c("ATGGGCT", "..TG--ACCGT..", "..GGuatgc..")

  dataset$add_seqs(names, sequences)

  # set total abundance. seq1 -> 100, seq2 -> 50, seq3 -> 5
  abunds <- list(100, 50, 5)
  dataset$set_abundances(names, abunds)

  # add group data
  groups <- c("sample1", "sample1", "sample2")
  dataset$set_group_assignments(names, groups)

  # set the abundance of 'seq1' in 'sample1' to 150
  dataset$set_abundances(c("seq1"), list(c(150, 0)))

  # set the abundance of 'seq1' in 'sample1' to 150,
  #                      'seq1' in 'sample2' to 0 and
  # set the abundance of 'seq3' in 'sample1' to 75,
  #                      'seq3' in 'sample2' to 5
  dataset$set_abundances(c("seq1", "seq3"), list(c(150, 0), c(75, 5)))

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data_vector-set_seqs"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data_vector-set_seqs}{}}}
\subsection{Method \code{set_seqs()}}{
Set existing sequence data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data_vector$set_seqs(
  names,
  sequences,
  comments = c(),
  search_scores = c(),
  sim_scores = c(),
  longest_inserts = c()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{names}}{a vector of sequence names}

\item{\code{sequences}}{a vector of sequence data}

\item{\code{comments}}{a vector of sequence comments, (optional)}

\item{\code{search_scores}}{a vector of alignment search scores,
(optional, used by align_seqs)}

\item{\code{sim_scores}}{a vector of alignment similarity to template scores,
(optional, used by align_seqs)}

\item{\code{longest_inserts}}{a vector of longest inserts in the alignment,
(optional, used by align_seqs)}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  dataset <- sequence_data_vector$new()
  names <- c("seq1", "seq2", "seq3")
  sequences <- c("ATGGGCT", "..TG--ACCGT..", "..GGuatgc..")
  dataset$add_seqs(names, sequences)

  # change sequence string for seq2 from ..TG--ACCGT.. to ...ATG-G-GCT..
  dataset$set_seqs(c("seq2"), c("...ATG-G-GCT.."))

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data_vector-write_count_file"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data_vector-write_count_file}{}}}
\subsection{Method \code{write_count_file()}}{
Write count table data to mothur formatted count file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data_vector$write_count_file(filename, path = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{String, name of file you want to write to}

\item{\code{path}}{path to count file (optional)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_data_vector-clone"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_data_vector-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_data_vector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
