% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r6_sequence-dataset.R
\name{sequence_dataset}
\alias{sequence_dataset}
\title{sequence_dataset}
\description{
'sequence_dataset' is an R6 class that represents sequence FASTA
and abundance data.
}
\examples{

## ------------------------------------------------
## Method `sequence_dataset$new`
## ------------------------------------------------

  dataset <- sequence_data_vector$new()

## ------------------------------------------------
## Method `sequence_dataset$read_fasta_file`
## ------------------------------------------------

  dataset <- sequence_dataset$new()
  fasta_data <- dataset$read_fasta_file(filename =
  rchime_example("test.fasta"))

## ------------------------------------------------
## Method `sequence_dataset$write_fasta_file`
## ------------------------------------------------

  dataset <- sequence_dataset$new()
  fasta_data <- dataset$read_fasta_file(filename =
  rchime_example("test.fasta"))
 dataset$write_fasta_file(fasta_data$names,
  fasta_data$sequences, filename = "output.fasta")
 remove_file("output.fasta")
}
\author{
Sarah Westcott, \email{swestcot@umich.edu}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-sequence_dataset-new}{\code{sequence_dataset$new()}}
\item \href{#method-sequence_dataset-abort_name_missing}{\code{sequence_dataset$abort_name_missing()}}
\item \href{#method-sequence_dataset-abort_length_mismatch}{\code{sequence_dataset$abort_length_mismatch()}}
\item \href{#method-sequence_dataset-alert_missing_name}{\code{sequence_dataset$alert_missing_name()}}
\item \href{#method-sequence_dataset-alert_eliminated_name}{\code{sequence_dataset$alert_eliminated_name()}}
\item \href{#method-sequence_dataset-calc_ambigs}{\code{sequence_dataset$calc_ambigs()}}
\item \href{#method-sequence_dataset-calc_longest_polymer}{\code{sequence_dataset$calc_longest_polymer()}}
\item \href{#method-sequence_dataset-calc_numbases}{\code{sequence_dataset$calc_numbases()}}
\item \href{#method-sequence_dataset-calc_numns}{\code{sequence_dataset$calc_numns()}}
\item \href{#method-sequence_dataset-calc_start}{\code{sequence_dataset$calc_start()}}
\item \href{#method-sequence_dataset-calc_end}{\code{sequence_dataset$calc_end()}}
\item \href{#method-sequence_dataset-read_fasta_file}{\code{sequence_dataset$read_fasta_file()}}
\item \href{#method-sequence_dataset-write_fasta_file}{\code{sequence_dataset$write_fasta_file()}}
\item \href{#method-sequence_dataset-clone}{\code{sequence_dataset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_dataset-new"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_dataset-new}{}}}
\subsection{Method \code{new()}}{
Create a new FASTA sequence dataset parent object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_dataset$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A new \code{sequence_data_vector} object which inherits
from sequence_dataset
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  dataset <- sequence_data_vector$new()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_dataset-abort_name_missing"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_dataset-abort_name_missing}{}}}
\subsection{Method \code{abort_name_missing()}}{
Report name missing and abort
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_dataset$abort_name_missing(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{String, containing name of missing sequence}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_dataset-abort_length_mismatch"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_dataset-abort_length_mismatch}{}}}
\subsection{Method \code{abort_length_mismatch()}}{
Report length mismatch and abort
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_dataset$abort_length_mismatch(name1, name2, length1, length2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name1}}{String, containing name of first variable}

\item{\code{name2}}{String, containing name of second variable}

\item{\code{length1}}{Numeric, containing length of first variable}

\item{\code{length2}}{Numeric, containing length of second variable}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_dataset-alert_missing_name"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_dataset-alert_missing_name}{}}}
\subsection{Method \code{alert_missing_name()}}{
Report missing name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_dataset$alert_missing_name(name, return_value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{String, containing name of missing sequence}

\item{\code{return_value}}{String, containing return_value}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_dataset-alert_eliminated_name"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_dataset-alert_eliminated_name}{}}}
\subsection{Method \code{alert_eliminated_name()}}{
Report eliminated name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_dataset$alert_eliminated_name(name, return_value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{String, containing name of eliminated sequence}

\item{\code{return_value}}{String, containing return_value}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_dataset-calc_ambigs"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_dataset-calc_ambigs}{}}}
\subsection{Method \code{calc_ambigs()}}{
Get number of ambiguous bases
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_dataset$calc_ambigs(seq)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seq}}{String, containing nucleotides}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An Integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_dataset-calc_longest_polymer"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_dataset-calc_longest_polymer}{}}}
\subsection{Method \code{calc_longest_polymer()}}{
Get length of longest homopolymer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_dataset$calc_longest_polymer(seq)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seq}}{String, containing nucleotides}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An Integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_dataset-calc_numbases"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_dataset-calc_numbases}{}}}
\subsection{Method \code{calc_numbases()}}{
Get number of bases in sequence
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_dataset$calc_numbases(seq)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seq}}{String, containing nucleotides}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An Integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_dataset-calc_numns"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_dataset-calc_numns}{}}}
\subsection{Method \code{calc_numns()}}{
Get number of 'N' bases
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_dataset$calc_numns(seq)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seq}}{String, containing nucleotides}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An Integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_dataset-calc_start"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_dataset-calc_start}{}}}
\subsection{Method \code{calc_start()}}{
Get location of first base in sequence
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_dataset$calc_start(seq)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seq}}{String, containing nucleotides}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An Integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_dataset-calc_end"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_dataset-calc_end}{}}}
\subsection{Method \code{calc_end()}}{
Get location of last base in sequence
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_dataset$calc_end(seq)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seq}}{String, containing nucleotides}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An Integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_dataset-read_fasta_file"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_dataset-read_fasta_file}{}}}
\subsection{Method \code{read_fasta_file()}}{
Read a FASTA formatted sequence file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_dataset$read_fasta_file(filename, path = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{FASTA file name (required)}

\item{\code{path}}{path to FASTA file (optional)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data.frame containing the FASTA sequence data and summary
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  dataset <- sequence_dataset$new()
  fasta_data <- dataset$read_fasta_file(filename =
  rchime_example("test.fasta"))
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_dataset-write_fasta_file"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_dataset-write_fasta_file}{}}}
\subsection{Method \code{write_fasta_file()}}{
Write dataset to FASTA file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_dataset$write_fasta_file(names, sequences, filename)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{names}}{vector of strings containing sequence names}

\item{\code{sequences}}{vector of strings containing sequence nucleotides}

\item{\code{filename}}{String, name of file you want to write FASTA
sequences to}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{  dataset <- sequence_dataset$new()
  fasta_data <- dataset$read_fasta_file(filename =
  rchime_example("test.fasta"))
 dataset$write_fasta_file(fasta_data$names,
  fasta_data$sequences, filename = "output.fasta")
 remove_file("output.fasta")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_dataset-clone"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_dataset-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_dataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
