% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silva_gold.R
\name{silva_gold}
\alias{silva_gold}
\title{silva_gold}
\usage{
silva_gold()
}
\value{
data.frame containing
\href{https://mothur.org/wiki/silva_reference_files/}{silva reference files
for chimera detection}
}
\description{
The \code{silva_gold()} function creates a silva based reference for use with the
\code{rchime()} function in reference mode. It contains 5,181 sequences.
}
\examples{

reference <- silva_gold()

}
\author{
Sarah Westcott, \email{swestcot@umich.edu}
}
