% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_strollur_objects.R
\name{strollur_multi_sample_small}
\alias{strollur_multi_sample_small}
\title{Create a
\href{https://mothur.org/strollur/reference/strollur.html}{strollur object}
object with 500 sequences}
\usage{
strollur_multi_sample_small()
}
\value{
\href{https://mothur.org/strollur/reference/strollur.html}{strollur
object} with 500 sequences
}
\description{
Create a \href{https://mothur.org/strollur/reference/strollur.html}{strollur
object} object with 500 sequences
}
\examples{
rchime::strollur_multi_sample_small()
}
