% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{summarize_fasta}
\alias{summarize_fasta}
\title{summarize_fasta}
\usage{
summarize_fasta(summary_report, count, processors)
}
\arguments{
\item{summary_report}{DataFrame, containing fasta summary data}

\item{count}{NumericalVector, containing abundances of sequences}

\item{processors}{Integer, number of cores to use. Default = all available}
}
\value{
DataFrame with summary values
}
\description{
Summarizes a fasta dataset
}
