% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_seqs.R
\name{summary_seqs}
\alias{summary_seqs}
\title{summary_seqs options}
\usage{
summary_seqs(
  dataset = NULL,
  processors = parallelly::availableCores(),
  silent = FALSE
)
}
\arguments{
\item{dataset}{R6 'sequence_dataset' object containing sequence data}

\item{processors}{Integer, number of cores to use. Default = all available}

\item{silent}{Boolean, suppress progress updates and console output}
}
\value{
list of data.tables containing the summary data
}
\description{
summary_seqs options
}
\examples{

# Let's create a dataset to summarize

dataset <- sequence_data_table$new(
  filename = mothur2_example("test.fasta")
)
dataset$set_group_assignments(
  filename = mothur2_example("test.count_table")
)

# To summarize the dataset, run the following:

summary <- summary_seqs(dataset = dataset$dataset)

}
\author{
Sarah Westcott, \email{swestcot@umich.edu}
}
