% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uchime_help.R
\name{uchimeout_help}
\alias{uchimeout_help}
\title{uchimeout_help}
\usage{
uchimeout_help()
}
\description{
Description of 'uchimeout' data.frame returned by the
[\code{\link[=rchime]{rchime()}}] function.
}
\author{
Sarah Westcott, \email{swestcot@umich.edu}
}
