% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chimera_uchime.R
\name{chimera_uchime}
\alias{chimera_uchime}
\title{chimera_uchime}
\usage{
chimera_uchime(
  dataset = NULL,
  fasta = NULL,
  count = NULL,
  reference = NULL,
  dereplicate = FALSE,
  processors = parallelly::availableCores(),
  silent = FALSE,
  chimealns = FALSE,
  abskew = 2,
  minh = 0.3,
  mindiv = 0.5,
  xn = 8,
  dn = 1.4,
  xa = 1,
  chunks = 4,
  minchunk = 64,
  idsmoothwindow = 32,
  maxp = 2,
  skipgaps = TRUE,
  skipgaps2 = TRUE,
  minlen = 10,
  maxlen = 10000
)
}
\arguments{
\item{dataset}{R6 'sequence_dataset' object containing sequence data}

\item{fasta}{String, name of file containing FASTA reads}

\item{count}{String, name of mothur formatted
\href{https://mothur.org/wiki/count_file/}{count} file.}

\item{reference}{String, name of optional FASTA reference file.}

\item{dereplicate, }{Boolean, The dereplicate option allows you to remove
chimeras by sample. For example, if dereplicate parameter is false, then if
one group finds the sequence to be chimeric, it will be removed from all
groups. If dereplicate is set to true, sequences found to be chimeric are
only removed from the sample they are found to be chimeric in.
Default = FALSE.}

\item{processors}{Integer, number of cores to use. Default = all available}

\item{silent}{Boolean, suppress progress updates and console output
\subsection{Parameters used directly by uchime. It is recommended to use the default}{

values for the parameters below unless you have a pressing reason to change
them.
}}

\item{chimealns}{bool, Default = FALSE. When TRUE a DataFrame is returned
containing multiple alignments of query sequences to parents
in human-readable format. Alignments show columns with
differences that support or contradict a chimeric model.}

\item{abskew}{Float, the minimum abundance skew (denovo only). Default = 2.0.
abskew <- min (abund(parent1), abund(parent2)) / abund(query)}

\item{minh}{Float, Mininum score to report chimera. Default 0.3. Values from
0.1 to 5 might be reasonable. Lower values increase sensitivity
but may report more false positives. If you decrease --xn,
you may need to increase --minh, and vice versa.}

\item{mindiv}{Float, Minimum divergence ratio, default 0.5. Div ratio is
100\%\% \%\%identity between query sequence and the closest candidate for
being a parent. If you don't care about very close chimeras,
then you could increase --mindiv to, say, 1.0 or 2.0, and
also decrease --min h, say to 0.1, to increase sensitivity.
How well this works will depend on your data. Best is to
tune parameters on a good benchmark.}

\item{xn}{Float, Weight of a no vote, also called the beta parameter.
Default 8.0. Decreasing this weight to around 3 or 4 may give better
performance on denoised data.}

\item{dn}{Float, Pseudo-count prior on number of no votes. Default 1.4.
Probably no good reason to change this unless you can retune to a good
benchmark for your data. Reasonable values are probably in the
range from 0.2 to 2.}

\item{xa}{Float, Weight of an abstain vote. Default 1. So far, results do not
seem to be very sensitive to this parameter, but if you have
a good training set might be worth trying. Reasonable values
might range from 0.1 to 2.}

\item{chunks}{Integer, Number of chunks to extract from the query sequence
when searching for parents. Default 4.}

\item{minchunk}{Integer, Minimum length of a chunk. Default 64.}

\item{idsmoothwindow}{Integer, Length of id smoothing window. Default 32.}

\item{maxp}{Integer, Maximum number of candidate parents to consider.
Default 2. In tests so far, increasing maxp gives only a very small
improvement in sensivity but tends to increase the error rate quite a bit.}

\item{skipgaps, }{bool, If skipgaps is specified, columns containing gaps do
not count as diffs. Default = TRUE.}

\item{skipgaps2}{bool, If skipgaps2 is specified, if column is immediately
adjacent to a column containing a gap, it is not counted as a diff.
Default = TRUE.}

\item{minlen}{Integer, Minimum sequence length. Default = 10.}

\item{maxlen}{Integer, Maximum sequence length. Default = 10000.}
}
\value{
DataFrame containing chimera results and an updated dataset with the
chimeras removed
}
\description{
The chimera_uchime function detects and removes chimeras from your data.

This function uses modified
\href{http://drive5.com/usearch/manual/uchime_algo.html}{uchime} source code
written by Robert C. Edgar and donated to the public domain.

Edgar,R.C., Haas,B.J., Clemente,J.C., Quince,C. and Knight,R. (2011),
UCHIME improves sensitivity and speed of chimera detection.
Bioinformatics 27:2194.
}
\examples{

# Let's create a dataset from the example files

dataset <- sequence_data_table$new(
  filename = rchime_example("test.fasta")
)
dataset$set_group_assignments(
  filename = rchime_example("test.count_table")
)

# detect and remove chimeras from the dataset using denovo approach
# (recommended)

chimera_report <- chimera_uchime(dataset, dereplicate = TRUE)

# or detect and remove chimeras using a reference

dataset <- sequence_data_table$new(
  filename = rchime_example("test.fasta")
)
dataset$set_group_assignments(
  filename = rchime_example("test.count_table")
)
reference <- sequence_data_table$new(
  filename = rchime_example("reference.fasta")
)

chimera_report <- chimera_uchime(dataset, reference = reference)

}
\seealso{
\code{\link[=uchimeout_help]{uchimeout_help()}} for a description of the DataFrame returned
}
\author{
Sarah Westcott, \email{swestcot@umich.edu}
}
