% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rchime-package.R
\docType{package}
\name{rchime-package}
\alias{rchime-package}
\title{rchime: Detect and Remove Chimeras from Amplicon Sequence Analysis Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Detect and remove chimeras from your Amplicon Sequence Analysis using reference based or denovo approaches.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mothur/rchime}
  \item \url{https://mothur.org/rchime/}
  \item Report bugs at \url{https://github.com/mothur/rchime/issues}
}

}
\author{
\strong{Maintainer}: Sarah Westcott \email{swestcot@umich.edu} (\href{https://orcid.org/0009-0001-1529-8247}{ORCID})

Authors:
\itemize{
  \item Sarah Westcott \email{swestcot@umich.edu} (\href{https://orcid.org/0009-0001-1529-8247}{ORCID})
}

Other contributors:
\itemize{
  \item Pat Schloss \email{pschloss@umich.edu} (\href{https://orcid.org/0000-0002-6935-4275}{ORCID}) [copyright holder]
}

}
\keyword{internal}
