% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rchime.R
\name{rchime}
\alias{rchime}
\alias{rchime.strollur}
\alias{rchime.data.frame}
\title{rchime}
\usage{
rchime(
  data,
  reference = NULL,
  dereplicate = TRUE,
  verbose = TRUE,
  remove_chimeras = TRUE,
  rchime_options = NULL,
  table_names = list(sequence_name = "sequence_name", sequence = "sequence", abundance =
    "abundance", sample = "sample")
)

\method{rchime}{strollur}(
  data,
  reference = NULL,
  dereplicate = TRUE,
  verbose = TRUE,
  remove_chimeras = TRUE,
  rchime_options = NULL,
  table_names = list(sequence_name = "sequence_name", sequence = "sequence")
)

\method{rchime}{data.frame}(
  data,
  reference = NULL,
  dereplicate = TRUE,
  verbose = TRUE,
  remove_chimeras = TRUE,
  rchime_options = NULL,
  table_names = list(sequence_name = "sequence_name", sequence = "sequence", abundance =
    "abundance", sample = "sample")
)
}
\arguments{
\item{data}{a \href{https://mothur.org/strollur/}{strollur} dataset object or
a data.frame containing your sequence data.}

\item{reference}{a \href{https://mothur.org/strollur/}{strollur} dataset
object or a data.frame containing reference sequence data.}

\item{dereplicate}{logical. The dereplicate option allows you to remove
chimeras by sample. When \code{dereplicate=FALSE}, if a sequence is flagged as
chimeric in one sample, it is removed from all samples. Our experience
suggests that this is a bit aggressive since we’ve seen rare sequences get
flagged as chimeric when they’re the most abundant sequence in another
sample. For a more conservative approach, we recommend using the default
\code{dereplicate=TRUE} which will only remove sequences from the samples in
which they are flagged as chimeric.}

\item{verbose}{logical, allow console outputs. Default = \code{TRUE}.}

\item{remove_chimeras}{Boolean, remove chimeras from dataset. Default =
\code{TRUE}.}

\item{rchime_options}{List, You can fine tune the vsearch specific options
using  the [\code{\link[=rchime_options]{rchime_options()}}] function. Default = \code{NULL}.}

\item{table_names, }{named list used to indicate the names of the columns in
the data.frame. Only used when \code{data} or \code{reference} are a data.frames.
Default = table_names <- list(sequence_name = "sequence_name",
sequence = "sequence")

In table_names, 'sequence_name' is a string containing the name of the column
in 'table' that contains the sequence names. Default column name is
'sequence_name'.

In table_names, 'sequence' is a string containing the name of the
column in 'table' that contains the sequences. Default column name is
'sequence'.}
}
\value{
data a
\href{https://mothur.org/strollur/reference/strollur.html}{strollur object}
}
\description{
The \code{rchime()} function allows you to detect and remove chimeras from your
data using a de novo approach or alternatively a reference model.

Our preferred way of doing this is to use the abundant sequences as our
reference (de novo).

This function uses code from the
\href{https://github.com/torognes/vsearch}{vsearch} tools.
}
\examples{

# Let's use a strollur object with 500 sequences

data <- strollur_multi_sample_small()
rchime(data)
data


# Detect and remove chimeras from the dataset using de novo approach by
# sample (recommended)

table <- readRDS(rchime_example("miseq_data_frame_by_sample_small.rds"))

data <- rchime(data = table)
data

}
\references{
Rognes T, Flouri T, Nichols B, Quince C, Mahé F. (2016) VSEARCH: a
versatile open source tool for metagenomics. PeerJ 4:e2584.
doi: 10.7717/peerj.2584

Edgar,R.C., Haas,B.J., Clemente,J.C., Quince,C. and Knight,R. (2011),
UCHIME improves sensitivity and speed of chimera detection.
Bioinformatics 27:2194.
}
\seealso{
\code{\link[=rchime_options]{rchime_options()}} to set vsearch specific parameters.
}
\author{
Sarah Westcott, \email{swestcot@umich.edu}
}
