% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rchime.R
\name{rchime}
\alias{rchime}
\title{rchime}
\usage{
rchime(
  data,
  reference = NULL,
  dereplicate = FALSE,
  processors = parallelly::availableCores(),
  silent = FALSE,
  remove_chimeras = TRUE,
  rchime_options = NULL
)
}
\arguments{
\item{data}{a \href{https://mothur.org/strollur/}{strollur} dataset object
containing your sequence data.}

\item{reference}{a \href{https://mothur.org/strollur/}{strollur} dataset
object containing reference sequence data.}

\item{dereplicate, }{Boolean, The dereplicate option allows you to remove
chimeras by sample. For example, if dereplicate parameter is FALSE, then if
one group finds the sequence to be chimeric, it will be removed from all
groups. If dereplicate is set to TRUE, sequences found to be chimeric are
only removed from the sample they are found to be chimeric in.
Default = FALSE.}

\item{processors}{Integer, number of cores to use. Default = all available}

\item{silent}{Boolean, suppress concole outputs. Default = FALSE.}

\item{remove_chimeras}{Boolean, remove chimeras from dataset. Default = TRUE.}

\item{rchime_options}{List, You can fine tune the vsearch specific options
using  the [\code{\link[=rchime_options]{rchime_options()}}] function. Default = NULL.}
}
\value{
list() containing a chimera report, and vector of the chimeric
sequence's names.

The \href{https://mothur.org/strollur/}{strollur} dataset object will also be
updated. The sequences flagged as chimeric are removed and the chimera report
is added.
}
\description{
The rchime function allows you to remove chimeras from your data
using a denovo approach or alternatively a reference model.

Our preferred way of doing this is to use the abundant sequences
as our reference (denovo). In addition, if a sequence is flagged as chimeric
in one sample, the default (dereplicate=FALSE) is to remove it from
all samples. Our experience suggests that this is a bit aggressive
since we’ve seen rare sequences get flagged as chimeric when they’re
the most abundant sequence in another sample. For a more conservative
approach, set (dereplicate=TRUE) which will only remove sequences
from the samples in which they are flagged as chimeric.

This function uses code from the
\href{https://github.com/torognes/vsearch}{vsearch} tools.
}
\examples{

# Let's create a dataset named "rchime denovo example"

data_denovo <- strollur::new_dataset("rchime denovo example")

# Read the fasta and abundance data

fasta_data <- readRDS(rchime_example("miseq_fasta.rds"))
abundance_data <- readRDS(rchime_example("miseq_abundance.rds"))

# Add the sequence and abundance data to the dataset

strollur::add(data_denovo, table = fasta_data, type = "sequences")
strollur::assign(data_denovo,
  table = abundance_data,
  type = "sequence_abundance"
)

# Detect and remove chimeras from the dataset using denovo approach
# (recommended)

chimera_report <- rchime(data_denovo, dereplicate = TRUE)
data_denovo

# Alternatively you can detect and remove chimeras using a reference

data_reference <- strollur::new_dataset("rchime reference example")

strollur::add(data_reference, table = fasta_data, type = "sequences")
strollur::assign(data_reference,
  table = abundance_data,
  type = "sequence_abundance"
)

reference <- new_dataset("Silva V4 Region")

reference_data <- readRDS(rchime_example("reference.rds"))

strollur::add(reference, table = reference_data, type = "sequences")

chimera_report <- rchime(data_reference, reference = reference)
data_reference

}
\references{
Rognes T, Flouri T, Nichols B, Quince C, Mahé F. (2016) VSEARCH: a
versatile open source tool for metagenomics. PeerJ 4:e2584.
doi: 10.7717/peerj.2584

Edgar,R.C., Haas,B.J., Clemente,J.C., Quince,C. and Knight,R. (2011),
UCHIME improves sensitivity and speed of chimera detection.
Bioinformatics 27:2194.
}
\seealso{
\code{\link[=rchime_options]{rchime_options()}} to set vsearch specific parameters.
}
\author{
Sarah Westcott, \email{swestcot@umich.edu}
}
