% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rchime.R
\name{rchime}
\alias{rchime}
\title{rchime}
\usage{
rchime(
  dataset = NULL,
  fasta = NULL,
  count = NULL,
  reference = NULL,
  dereplicate = FALSE,
  processors = parallelly::availableCores(),
  silent = FALSE,
  chimealns = FALSE,
  rchime_options = NULL
)
}
\arguments{
\item{dataset}{R6 'sequence_dataset' object containing sequence data}

\item{fasta}{String, name of file containing FASTA reads}

\item{count}{String, name of mothur formatted
\href{https://mothur.org/wiki/count_file/}{count} file.}

\item{reference}{String, name of optional FASTA reference file.}

\item{dereplicate, }{Boolean, The dereplicate option allows you to remove
chimeras by sample. For example, if dereplicate parameter is false, then if
one group finds the sequence to be chimeric, it will be removed from all
groups. If dereplicate is set to true, sequences found to be chimeric are
only removed from the sample they are found to be chimeric in.
Default = FALSE.}

\item{processors}{Integer, number of cores to use. Default = all available}

\item{silent}{Boolean, suppress progress updates and console output}

\item{chimealns}{bool, Default = FALSE. When TRUE a DataFrame is returned
containing multiple alignments of query sequences to parents
in human-readable format. Alignments show columns with
differences that support or contradict a chimeric model.}

\item{rchime_options}{List, You can fine tune the uchime specific options
using  the [\code{\link[=rchime_options]{rchime_options()}}] function. Default = NULL.}
}
\value{
List containing accnos, uchimeout and optional chimealns.
}
\description{
The rchime function allows you to remove chimeras from your data
using a denovo approach or alternatively a reference model.

Our preferred way of doing this is to use the abundant sequences
as our reference (denovo). In addition, if a sequence is flagged as chimeric
in one sample, the default (dereplicate=FALSE) is to remove it from
all samples. Our experience suggests that this is a bit aggressive
since we’ve seen rare sequences get flagged as chimeric when they’re
the most abundant sequence in another sample. For a more conservative
approach, set (dereplicate=TRUE) which will only remove sequences
from the samples in which they are flagged as chimeric.

This function relies on a rewrite of Robert C. Edgar's
\href{http://drive5.com/usearch/manual/uchime_algo.html}{uchime} source code
which was donated to the public domain.
}
\examples{

# Let's create a dataset from the example files

dataset <- sequence_data_vector$new(
  filename = rchime_example("test.fasta")
)
dataset$set_group_assignments(
  filename = rchime_example("test.count_table")
)

# detect and remove chimeras from the dataset using denovo approach
# (recommended)

chimera_report <- rchime(dataset, dereplicate = TRUE)

# or detect and remove chimeras using a reference

dataset <- sequence_data_vector$new(
  filename = rchime_example("test.fasta")
)
dataset$set_group_assignments(
  filename = rchime_example("test.count_table")
)

chimera_report <- rchime(dataset,
  reference = rchime_example("reference.fasta")
)

}
\references{
Edgar,R.C., Haas,B.J., Clemente,J.C., Quince,C. and Knight,R. (2011),
UCHIME improves sensitivity and speed of chimera detection.
Bioinformatics 27:2194.
}
\seealso{
\code{\link[=uchimeout_help]{uchimeout_help()}} for a description of 'uchimeout'

\code{\link[=rchime_options]{rchime_options()}} to set uchime specific parameters.
}
\author{
Sarah Westcott, \email{swestcot@umich.edu}
}
