% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rchime.R
\name{rchime}
\alias{rchime}
\title{Detect and remove chimeras from your
\href{https://mothur.org/strollur/}{strollur} object or data.frame
using a denovo approach or alternatively a reference model.}
\usage{
rchime(
  data,
  reference = NULL,
  dereplicate = TRUE,
  silent = FALSE,
  remove_chimeras = TRUE,
  rchime_options = NULL,
  table_names = list(sequence_name = "sequence_name", sequence = "sequence", abundance =
    "abundance", sample = "sample")
)
}
\arguments{
\item{data}{a \href{https://mothur.org/strollur/}{strollur} dataset object or
a data.frame containing your sequence data.}

\item{reference}{a \href{https://mothur.org/strollur/}{strollur} dataset
object or a data.frame containing reference sequence data.}

\item{dereplicate}{logical. The dereplicate option allows you to remove
chimeras by sample. When \code{dereplicate=FALSE}, if a sequence is flagged as
chimeric in one sample, it is removed from all samples. Our experience
suggests that this is a bit aggressive since we’ve seen rare sequences get
flagged as chimeric when they’re the most abundant sequence in another
sample. For a more conservative approach, we recommend using the default
\code{dereplicate=TRUE} which will only remove sequences from the samples in
which they are flagged as chimeric.}

\item{silent}{Boolean, suppress console outputs. Default = FALSE.}

\item{remove_chimeras}{Boolean, remove chimeras from dataset. Default = TRUE.
Only used when \code{data} is a strollur object.}

\item{rchime_options}{List, You can fine tune the vsearch specific options
using  the [\code{\link[=rchime_options]{rchime_options()}}] function. Default = NULL.}

\item{table_names, }{named list used to indicate the names of the columns in
the data.frame. Only used when \code{data} or \code{reference} are a data.frames.}
}
\value{
list() containing a chimera report, and vector of the chimeric
sequence's names.

The \href{https://mothur.org/strollur/}{strollur} dataset object will also be
updated. The sequences flagged as chimeric are removed and the chimera report
is added.
}
\description{
The \code{rchime()} function allows you to detect and remove chimeras from your
data using a denovo approach or alternatively a reference model.

Our preferred way of doing this is to use the abundant sequences as our
reference (denovo).

This function uses code from the
\href{https://github.com/torognes/vsearch}{vsearch} tools.
}
\examples{

# Let's use a strollur object with 100 sequences

data <- strollur::load_dataset(
  rchime_example("strollur_multi_sample_tiny.rds")
)

chimera_report <- rchime(data)
data

}
\references{
Rognes T, Flouri T, Nichols B, Quince C, Mahé F. (2016) VSEARCH: a
versatile open source tool for metagenomics. PeerJ 4:e2584.
doi: 10.7717/peerj.2584

Edgar,R.C., Haas,B.J., Clemente,J.C., Quince,C. and Knight,R. (2011),
UCHIME improves sensitivity and speed of chimera detection.
Bioinformatics 27:2194.
}
\seealso{
\code{\link[=rchime_options]{rchime_options()}} to set vsearch specific parameters.
}
\author{
Sarah Westcott, \email{swestcot@umich.edu}
}
