% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rchime_options.R
\name{rchime_options}
\alias{rchime_options}
\title{rchime_options}
\usage{
rchime_options(
  abskew = 2,
  minh = 0.28,
  mindiv = 0.8,
  xn = 8,
  dn = 1.4,
  maxp = 3
)
}
\arguments{
\item{abskew}{Float, the minimum abundance skew (denovo only). Default = 2.0.
abskew <- min (abund(parent1), abund(parent2)) / abund(query)}

\item{minh}{Float, Mininum score to report chimera. Default 0.28. Values from
0.1 to 5 might be reasonable. Lower values increase sensitivity
but may report more false positives. If you decrease --xn,
you may need to increase --minh, and vice versa.}

\item{mindiv}{Float, Minimum divergence ratio, default 0.8. Div ratio is
100\%\% \%\%identity between query sequence and the closest candidate for
being a parent. If you don't care about very close chimeras,
then you could increase --mindiv to, say, 1.0 or 2.0, and
also decrease --min h, say to 0.1, to increase sensitivity.
How well this works will depend on your data. Best is to
tune parameters on a good benchmark.}

\item{xn}{Float, Weight of a no vote, also called the beta parameter.
Default 8.0. Decreasing this weight to around 3 or 4 may give better
performance on denoised data.}

\item{dn}{Float, Pseudo-count prior on number of no votes. Default 1.4.
Probably no good reason to change this unless you can retune to a good
benchmark for your data. Reasonable values are probably in the
range from 0.2 to 2.}

\item{maxp}{Integer, Maximum number of candidate parents to consider.
Default 3.}
}
\value{
List containing selected rchime parameters and their values.
}
\description{
The rchime_options function allows you to set parameters for the
[\code{\link[=rchime]{rchime()}}] function. We recommend using the defaults, unless you
have a compelling reason to do otherwise.
}
\references{
Rognes T, Flouri T, Nichols B, Quince C, Mahé F. (2016) VSEARCH: a
versatile open source tool for metagenomics. PeerJ 4:e2584.
doi: 10.7717/peerj.2584

Edgar,R.C., Haas,B.J., Clemente,J.C., Quince,C. and Knight,R. (2011),
UCHIME improves sensitivity and speed of chimera detection.
Bioinformatics 27:2194.
}
\author{
Sarah Westcott, \email{swestcot@umich.edu}
}
