% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rchimeDenovo}
\alias{rchimeDenovo}
\title{rchimeDenovo}
\usage{
rchimeDenovo(sequence_names, sequences, abundances, options)
}
\arguments{
\item{sequence_names, }{2D vector of strings containing sequence names parsed by sample}

\item{sequences, }{2D vector of strings containing sequence nucleotide data  parsed by sample}

\item{abundances, }{2D vector of sequence abundances parsed by sample}

\item{options}{list containing parameter options}
}
\description{
detects chimeras from your data using a denovo method processing by sample.
}
\seealso{
\code{\link[=rchime]{rchime()}}
}
