% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rchimeDenovo}
\alias{rchimeDenovo}
\title{Detects chimeras using a denovo method processing by sample.}
\usage{
rchimeDenovo(sequence_name, sequence, abundance, options = NULL)
}
\arguments{
\item{sequence_name, }{2D vector of strings containing sequence names parsed by sample}

\item{sequence, }{2D vector of strings containing sequence nucleotide data  parsed by sample}

\item{abundance, }{2D vector of sequence abundances parsed by sample}

\item{options}{list containing parameter options. Default = NULL.}
}
\value{
list()
}
\description{
detects chimeras from your data using a denovo method processing by sample.
}
\examples{

# Read in data.frames containing sequence and abundance data

sequence_name <- readRDS(rchime_example("miseq_names_by_sample.rds"))
sequence <- readRDS(rchime_example("miseq_sequences_by_sample.rds"))
abundance <- readRDS(rchime_example("miseq_abundance_by_sample.rds"))

# Detect chimeras

results <- rchimeDenovo(sequence_name = sequence_name,
                        sequence = sequence,
                        abundance = abundance)

}
\seealso{
\code{\link[=rchime]{rchime()}}
}
