% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rchimeDenovoSingleSample}
\alias{rchimeDenovoSingleSample}
\title{rchimeDenovoSingleSample}
\usage{
rchimeDenovoSingleSample(sequence_names, sequences, abundances, options)
}
\arguments{
\item{sequence_names, }{vector of strings containing sequence names}

\item{sequences, }{vector of strings containing sequence nucleotide data}

\item{abundances, }{vector of sequence abundances}

\item{options}{list containing parameter options}
}
\description{
detects chimeras from your data using a denovo method.
}
\seealso{
\code{\link[=rchime]{rchime()}}
}
