% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rchimeDenovoSingleSample}
\alias{rchimeDenovoSingleSample}
\title{Detects chimeras using a denovo approach}
\usage{
rchimeDenovoSingleSample(sequence_name, sequence, abundance, options = NULL)
}
\arguments{
\item{sequence_name, }{vector of strings containing sequence names}

\item{sequence, }{vector of strings containing sequence nucleotide data}

\item{abundance, }{vector of sequence abundances}

\item{options}{list containing parameter options. Default = NULL.}
}
\description{
detects chimeras in your data using a denovo approach.
}
\examples{

# Read in data.frames containing sequence and abundance data

fasta_data <- readRDS(rchime_example("miseq_fasta.rds"))
abundance_data <- readRDS(rchime_example("single_sample_abundance.rds"))

# Detect chimeras in first 100 sequences

results <- rchimeDenovoSingleSample(sequence_name = fasta_data$sequence_name[1:100],
                                    sequence = fasta_data$sequence[1:100],
                                    abundance = abundance_data$abundance[1:100])

}
\seealso{
\code{\link[=rchime]{rchime()}}
}
