% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rchimeReference}
\alias{rchimeReference}
\title{Detects chimeras using a reference based approach}
\usage{
rchimeReference(
  sequence_name,
  sequence,
  reference_name,
  reference_sequence,
  options = NULL
)
}
\arguments{
\item{sequence_name, }{vector of strings containing sequence names}

\item{sequence, }{vector of strings containing sequence nucleotide data}

\item{reference_name, }{vector of strings containing reference sequences names}

\item{reference_sequence, }{vector of strings containing reference sequences}

\item{options}{list containing parameter options. Default = NULL.}
}
\description{
detects chimeras in your data using a reference based approach.
}
\examples{

# Read in data.frames containing sequence data

fasta_data <- readRDS(rchime_example("miseq_fasta.rds"))
reference_data <- readRDS(rchime_example("reference.rds"))

# Detect chimeras in first 100 sequences

results <- rchimeReference(sequence_name = fasta_data$sequence_name[1:100],
                           sequence = fasta_data$sequence[1:100],
                           reference_name = reference_data$sequence_name,
                           reference_sequence = reference_data$sequence)

}
\seealso{
\code{\link[=rchime]{rchime()}}
}
